Remove the TrackMate style element classes and resources.
We moved everything to the config-ui artifact.
