+++ src/main/java/fiji/plugin/trackmate/visualization/PerEdgeFeatureColorGenerator.java
@@ -24,9 +24,9 @@ package fiji.plugin.trackmate.visualization;
 import java.awt.Color;
 
 import org.jgrapht.graph.DefaultWeightedEdge;
+import org.scijava.ui.config.visitors.gui.elements.colormap.Colormap;
 
 import fiji.plugin.trackmate.Model;
-import fiji.plugin.trackmate.gui.displaysettings.Colormap;
 
 public class PerEdgeFeatureColorGenerator implements TrackColorGenerator
 {
