+++ src/main/java/fiji/plugin/trackmate/visualization/SpotColorGeneratorPerEdgeFeature.java
@@ -25,10 +25,10 @@ import java.awt.Color;
 import java.util.Set;
 
 import org.jgrapht.graph.DefaultWeightedEdge;
+import org.scijava.ui.config.visitors.gui.elements.colormap.Colormap;
 
 import fiji.plugin.trackmate.Model;
 import fiji.plugin.trackmate.Spot;
-import fiji.plugin.trackmate.gui.displaysettings.Colormap;
 
 public class SpotColorGeneratorPerEdgeFeature implements FeatureColorGenerator< Spot >
 {
