+++ src/main/java/fiji/plugin/trackmate/visualization/SpotColorGeneratorPerTrackFeature.java
@@ -23,9 +23,10 @@ package fiji.plugin.trackmate.visualization;
 
 import java.awt.Color;
 
+import org.scijava.ui.config.visitors.gui.elements.colormap.Colormap;
+
 import fiji.plugin.trackmate.Model;
 import fiji.plugin.trackmate.Spot;
-import fiji.plugin.trackmate.gui.displaysettings.Colormap;
 
 public class SpotColorGeneratorPerTrackFeature implements FeatureColorGenerator< Spot >
 {
