--- src/main/resources/fiji/plugin/trackmate/gui/displaysettings/luts/Algae.lut
@@ -1,257 +1,0 @@
-Index	Red	Green	Blue
-0	17	36	20
-1	17	37	20
-2	18	37	21
-3	18	38	22
-4	18	39	22
-5	19	40	23
-6	19	41	24
-7	19	42	25
-8	20	43	25
-9	20	43	26
-10	20	44	27
-11	21	45	27
-12	21	46	28
-13	21	47	29
-14	21	48	29
-15	22	49	30
-16	22	49	31
-17	22	50	31
-18	22	51	32
-19	23	52	33
-20	23	53	33
-21	23	54	34
-22	23	55	35
-23	23	55	35
-24	23	56	36
-25	24	57	36
-26	24	58	37
-27	24	59	38
-28	24	60	38
-29	24	61	39
-30	24	61	40
-31	25	62	40
-32	25	63	41
-33	25	64	41
-34	25	65	42
-35	25	66	43
-36	25	67	43
-37	25	67	44
-38	25	68	44
-39	25	69	45
-40	25	70	46
-41	25	71	46
-42	25	72	47
-43	25	73	47
-44	25	73	48
-45	25	74	49
-46	25	75	49
-47	25	76	50
-48	25	77	50
-49	25	78	51
-50	25	79	51
-51	25	80	52
-52	25	80	52
-53	25	81	53
-54	25	82	54
-55	25	83	54
-56	25	84	55
-57	25	85	55
-58	25	86	56
-59	25	87	56
-60	24	87	57
-61	24	88	57
-62	24	89	58
-63	24	90	58
-64	24	91	59
-65	23	92	59
-66	23	93	60
-67	23	94	60
-68	23	94	61
-69	23	95	61
-70	22	96	62
-71	22	97	62
-72	22	98	63
-73	22	99	63
-74	21	100	63
-75	21	101	64
-76	21	102	64
-77	20	102	65
-78	20	103	65
-79	19	104	66
-80	19	105	66
-81	19	106	67
-82	18	107	67
-83	18	108	67
-84	18	109	68
-85	17	110	68
-86	17	110	69
-87	16	111	69
-88	16	112	69
-89	15	113	70
-90	15	114	70
-91	14	115	70
-92	14	116	71
-93	13	117	71
-94	13	118	72
-95	12	119	72
-96	12	119	72
-97	11	120	73
-98	11	121	73
-99	10	122	73
-100	10	123	74
-101	9	124	74
-102	9	125	74
-103	8	126	74
-104	8	127	75
-105	7	128	75
-106	7	129	75
-107	7	129	76
-108	7	130	76
-109	6	131	76
-110	6	132	76
-111	6	133	77
-112	6	134	77
-113	6	135	77
-114	7	136	77
-115	7	137	77
-116	7	138	78
-117	8	138	78
-118	8	139	78
-119	9	140	78
-120	9	141	78
-121	10	142	79
-122	11	143	79
-123	12	144	79
-124	13	145	79
-125	15	146	79
-126	16	146	79
-127	17	147	79
-128	19	148	80
-129	20	149	80
-130	21	150	80
-131	23	151	80
-132	25	152	80
-133	26	153	80
-134	28	153	80
-135	30	154	80
-136	31	155	80
-137	33	156	81
-138	35	157	81
-139	37	157	81
-140	39	158	81
-141	41	159	81
-142	43	160	81
-143	45	161	81
-144	47	161	82
-145	50	162	82
-146	52	163	82
-147	54	164	82
-148	56	164	83
-149	58	165	83
-150	61	166	84
-151	63	167	84
-152	65	167	84
-153	67	168	85
-154	69	169	86
-155	71	169	86
-156	74	170	87
-157	76	171	87
-158	78	171	88
-159	80	172	89
-160	82	173	90
-161	83	174	90
-162	85	174	91
-163	87	175	92
-164	89	176	93
-165	91	176	94
-166	93	177	95
-167	94	178	96
-168	96	178	96
-169	98	179	97
-170	100	180	98
-171	101	180	99
-172	103	181	100
-173	104	182	101
-174	106	183	102
-175	108	183	103
-176	109	184	104
-177	111	185	105
-178	112	185	107
-179	114	186	108
-180	115	187	109
-181	117	188	110
-182	118	188	111
-183	120	189	112
-184	121	190	113
-185	123	191	114
-186	124	191	115
-187	126	192	117
-188	127	193	118
-189	129	194	119
-190	130	194	120
-191	132	195	121
-192	133	196	122
-193	134	197	124
-194	136	197	125
-195	137	198	126
-196	139	199	127
-197	140	200	128
-198	141	200	130
-199	143	201	131
-200	144	202	132
-201	145	203	133
-202	147	203	135
-203	148	204	136
-204	149	205	137
-205	151	206	138
-206	152	207	140
-207	153	207	141
-208	155	208	142
-209	156	209	143
-210	157	210	145
-211	159	211	146
-212	160	211	147
-213	161	212	149
-214	163	213	150
-215	164	214	151
-216	165	215	153
-217	166	215	154
-218	168	216	155
-219	169	217	156
-220	170	218	158
-221	172	219	159
-222	173	219	160
-223	174	220	162
-224	175	221	163
-225	177	222	165
-226	178	223	166
-227	179	224	167
-228	180	225	169
-229	182	225	170
-230	183	226	171
-231	184	227	173
-232	186	228	174
-233	187	229	176
-234	188	230	177
-235	189	231	178
-236	191	231	180
-237	192	232	181
-238	193	233	183
-239	194	234	184
-240	196	235	185
-241	197	236	187
-242	198	237	188
-243	199	238	190
-244	201	239	191
-245	202	239	193
-246	203	240	194
-247	204	241	195
-248	206	242	197
-249	207	243	198
-250	208	244	200
-251	209	245	201
-252	211	246	203
-253	212	247	204
-254	213	248	206
-255	214	249	207
