--- src/main/resources/fiji/plugin/trackmate/gui/displaysettings/luts/Amp.lut
@@ -1,257 +1,0 @@
-Index	Red	Green	Blue
-0	60	9	17
-1	61	9	18
-2	62	9	19
-3	63	9	19
-4	65	10	20
-5	66	10	20
-6	67	10	21
-7	69	10	22
-8	70	11	22
-9	71	11	23
-10	72	11	24
-11	74	11	24
-12	75	12	25
-13	76	12	25
-14	77	12	26
-15	79	12	27
-16	80	12	27
-17	81	13	28
-18	83	13	28
-19	84	13	29
-20	85	13	29
-21	86	13	30
-22	88	13	30
-23	89	13	31
-24	90	14	32
-25	92	14	32
-26	93	14	32
-27	94	14	33
-28	96	14	33
-29	97	14	34
-30	98	14	34
-31	100	14	35
-32	101	14	35
-33	102	14	36
-34	104	14	36
-35	105	14	37
-36	106	14	37
-37	108	14	37
-38	109	14	38
-39	110	14	38
-40	112	14	38
-41	113	14	39
-42	114	14	39
-43	116	14	39
-44	117	14	39
-45	118	14	40
-46	120	14	40
-47	121	14	40
-48	122	14	40
-49	124	14	40
-50	125	13	41
-51	126	13	41
-52	128	13	41
-53	129	14	41
-54	130	14	41
-55	132	14	41
-56	133	14	41
-57	134	14	41
-58	135	14	41
-59	137	14	41
-60	138	14	41
-61	139	15	41
-62	141	15	40
-63	142	16	40
-64	143	16	40
-65	144	16	40
-66	145	17	40
-67	147	18	40
-68	148	18	40
-69	149	19	39
-70	150	19	39
-71	151	20	39
-72	152	21	39
-73	154	22	38
-74	155	23	38
-75	156	24	38
-76	157	24	38
-77	158	25	38
-78	159	26	37
-79	160	27	37
-80	161	28	37
-81	162	30	37
-82	163	31	37
-83	164	32	36
-84	165	33	36
-85	166	34	36
-86	166	35	36
-87	167	36	36
-88	168	37	36
-89	169	39	36
-90	170	40	36
-91	171	41	36
-92	172	42	36
-93	172	44	36
-94	173	45	36
-95	174	46	36
-96	175	47	36
-97	175	49	36
-98	176	50	36
-99	177	51	37
-100	177	53	37
-101	178	54	37
-102	179	55	38
-103	179	57	38
-104	180	58	38
-105	180	59	39
-106	181	61	40
-107	181	62	40
-108	182	63	41
-109	183	65	41
-110	183	66	42
-111	184	67	43
-112	184	69	44
-113	185	70	44
-114	185	71	45
-115	186	73	46
-116	186	74	47
-117	187	75	48
-118	187	76	49
-119	188	78	50
-120	188	79	51
-121	188	80	52
-122	189	81	53
-123	189	83	54
-124	190	84	55
-125	190	85	56
-126	191	86	57
-127	191	88	58
-128	192	89	60
-129	192	90	61
-130	192	91	62
-131	193	93	63
-132	193	94	64
-133	194	95	65
-134	194	96	67
-135	194	97	68
-136	195	99	69
-137	195	100	70
-138	196	101	71
-139	196	102	73
-140	196	103	74
-141	197	105	75
-142	197	106	77
-143	198	107	78
-144	198	108	79
-145	198	109	80
-146	199	110	82
-147	199	112	83
-148	199	113	84
-149	200	114	86
-150	200	115	87
-151	201	116	88
-152	201	117	90
-153	201	119	91
-154	202	120	92
-155	202	121	94
-156	202	122	95
-157	203	123	96
-158	203	124	98
-159	204	125	99
-160	204	127	100
-161	204	128	102
-162	205	129	103
-163	205	130	105
-164	205	131	106
-165	206	132	107
-166	206	133	109
-167	206	135	110
-168	207	136	111
-169	207	137	113
-170	207	138	114
-171	208	139	116
-172	208	140	117
-173	208	141	119
-174	209	143	120
-175	209	144	121
-176	209	145	123
-177	210	146	124
-178	210	147	126
-179	211	148	127
-180	211	149	128
-181	211	151	130
-182	212	152	131
-183	212	153	133
-184	212	154	134
-185	213	155	136
-186	213	156	137
-187	213	157	138
-188	214	158	140
-189	214	160	141
-190	214	161	143
-191	215	162	144
-192	215	163	146
-193	215	164	147
-194	216	165	148
-195	216	166	150
-196	216	168	151
-197	217	169	153
-198	217	170	154
-199	217	171	156
-200	218	172	157
-201	218	173	159
-202	218	174	160
-203	219	176	161
-204	219	177	163
-205	219	178	164
-206	220	179	166
-207	220	180	167
-208	220	181	169
-209	221	182	170
-210	221	184	172
-211	222	185	173
-212	222	186	174
-213	222	187	176
-214	223	188	177
-215	223	189	179
-216	223	191	180
-217	224	192	182
-218	224	193	183
-219	224	194	185
-220	225	195	186
-221	225	196	188
-222	226	198	189
-223	226	199	190
-224	226	200	192
-225	227	201	193
-226	227	202	195
-227	228	203	196
-228	228	204	198
-229	228	206	199
-230	229	207	200
-231	229	208	202
-232	230	209	203
-233	230	210	205
-234	230	212	206
-235	231	213	208
-236	231	214	209
-237	232	215	211
-238	232	216	212
-239	233	217	213
-240	233	219	215
-241	233	220	216
-242	234	221	218
-243	234	222	219
-244	235	223	220
-245	235	225	222
-246	236	226	223
-247	236	227	225
-248	237	228	226
-249	237	229	227
-250	238	231	229
-251	238	232	230
-252	239	233	232
-253	240	234	233
-254	240	235	234
-255	241	236	236
