--- src/main/resources/fiji/plugin/trackmate/gui/displaysettings/luts/Balance.lut
@@ -1,257 +1,0 @@
-Index	Red	Green	Blue
-0	23	28	66
-1	24	29	69
-2	25	31	72
-3	26	32	75
-4	27	33	78
-5	28	35	81
-6	29	36	85
-7	30	37	88
-8	31	39	91
-9	32	40	94
-10	33	41	97
-11	34	43	101
-12	35	44	104
-13	35	45	107
-14	36	47	111
-15	37	48	114
-16	37	49	118
-17	38	51	121
-18	39	52	125
-19	39	53	128
-20	40	55	132
-21	40	56	136
-22	40	57	139
-23	41	58	143
-24	41	60	147
-25	41	61	150
-26	41	63	154
-27	40	64	158
-28	40	65	161
-29	39	67	165
-30	39	68	168
-31	38	70	172
-32	36	72	175
-33	35	74	178
-34	33	75	181
-35	30	77	184
-36	27	80	186
-37	24	82	187
-38	21	84	188
-39	18	86	189
-40	15	89	189
-41	13	91	190
-42	11	93	190
-43	10	95	189
-44	9	98	189
-45	10	100	189
-46	11	102	189
-47	12	104	188
-48	14	106	188
-49	16	108	188
-50	19	110	187
-51	21	112	187
-52	24	114	187
-53	27	116	187
-54	29	117	186
-55	32	119	186
-56	35	121	186
-57	37	123	186
-58	40	125	186
-59	43	127	186
-60	45	128	185
-61	48	130	185
-62	51	132	185
-63	53	134	185
-64	56	135	185
-65	59	137	185
-66	61	139	185
-67	64	140	185
-68	66	142	185
-69	69	144	185
-70	72	145	185
-71	74	147	186
-72	77	149	186
-73	80	150	186
-74	83	152	186
-75	86	154	186
-76	89	155	186
-77	91	157	186
-78	94	158	187
-79	97	160	187
-80	100	162	187
-81	104	163	188
-82	107	165	188
-83	110	166	188
-84	113	168	189
-85	116	169	189
-86	119	171	190
-87	123	172	190
-88	126	173	191
-89	129	175	191
-90	132	176	192
-91	135	178	193
-92	139	179	193
-93	142	181	194
-94	145	182	195
-95	148	184	196
-96	151	185	197
-97	154	186	197
-98	157	188	198
-99	160	189	199
-100	163	191	200
-101	166	192	201
-102	169	194	202
-103	172	195	203
-104	175	197	204
-105	178	198	206
-106	181	200	207
-107	184	202	208
-108	187	203	209
-109	190	205	210
-110	193	206	212
-111	196	208	213
-112	199	210	214
-113	202	211	215
-114	204	213	217
-115	207	214	218
-116	210	216	219
-117	213	218	221
-118	216	220	222
-119	219	221	223
-120	221	223	225
-121	224	225	226
-122	227	227	228
-123	230	228	229
-124	232	230	231
-125	235	232	232
-126	238	234	234
-127	240	236	235
-128	240	236	235
-129	239	233	232
-130	238	231	230
-131	237	229	227
-132	236	226	224
-133	235	224	221
-134	234	222	218
-135	233	219	215
-136	232	217	213
-137	231	214	210
-138	231	212	207
-139	230	210	204
-140	229	207	201
-141	228	205	198
-142	227	203	195
-143	227	200	192
-144	226	198	190
-145	225	196	187
-146	224	193	184
-147	223	191	181
-148	223	189	178
-149	222	186	175
-150	221	184	172
-151	221	182	169
-152	220	179	166
-153	219	177	163
-154	218	175	161
-155	218	173	158
-156	217	170	155
-157	216	168	152
-158	216	166	149
-159	215	163	146
-160	214	161	143
-161	214	159	140
-162	213	157	138
-163	212	154	135
-164	212	152	132
-165	211	150	129
-166	210	148	126
-167	210	145	123
-168	209	143	120
-169	208	141	118
-170	208	138	115
-171	207	136	112
-172	206	134	109
-173	205	132	106
-174	205	129	104
-175	204	127	101
-176	203	125	98
-177	203	122	95
-178	202	120	93
-179	201	118	90
-180	200	115	87
-181	200	113	85
-182	199	111	82
-183	198	108	79
-184	197	106	77
-185	197	104	74
-186	196	101	72
-187	195	99	69
-188	194	96	67
-189	193	94	64
-190	192	92	62
-191	192	89	60
-192	191	87	58
-193	190	84	55
-194	189	82	53
-195	188	79	51
-196	187	77	49
-197	186	74	47
-198	185	72	45
-199	184	69	44
-200	183	66	42
-201	182	64	41
-202	181	61	40
-203	180	58	39
-204	179	56	38
-205	177	53	37
-206	176	50	36
-207	175	48	36
-208	173	45	36
-209	172	43	36
-210	170	40	36
-211	168	38	36
-212	167	35	36
-213	165	33	36
-214	163	31	37
-215	161	29	37
-216	159	27	37
-217	157	25	38
-218	155	23	38
-219	153	21	39
-220	150	20	39
-221	148	18	40
-222	146	17	40
-223	143	16	40
-224	141	15	40
-225	138	15	41
-226	136	14	41
-227	133	14	41
-228	130	14	41
-229	128	13	41
-230	125	13	41
-231	122	14	40
-232	120	14	40
-233	117	14	39
-234	114	14	39
-235	112	14	38
-236	109	14	38
-237	106	14	37
-238	104	14	36
-239	101	14	35
-240	98	14	34
-241	96	14	33
-242	93	14	33
-243	90	14	32
-244	88	13	30
-245	85	13	29
-246	83	13	28
-247	80	12	27
-248	77	12	26
-249	75	12	25
-250	72	11	24
-251	70	11	22
-252	67	10	21
-253	65	10	20
-254	62	9	19
-255	60	9	17
