--- src/main/resources/fiji/plugin/trackmate/gui/displaysettings/luts/Curl.lut
@@ -1,257 +1,0 @@
-Index	Red	Green	Blue
-0	20	29	67
-1	21	31	68
-2	21	33	69
-3	22	35	70
-4	22	37	71
-5	23	39	72
-6	23	41	74
-7	24	43	75
-8	24	45	76
-9	24	47	77
-10	25	48	78
-11	25	50	79
-12	25	52	80
-13	26	54	82
-14	26	56	83
-15	26	58	84
-16	27	60	85
-17	27	61	86
-18	27	63	87
-19	27	65	89
-20	27	67	90
-21	27	69	91
-22	28	71	92
-23	28	72	93
-24	28	74	94
-25	28	76	96
-26	28	78	97
-27	27	80	98
-28	27	82	99
-29	27	83	100
-30	27	85	101
-31	27	87	102
-32	26	89	103
-33	26	91	105
-34	26	93	106
-35	25	95	107
-36	25	96	108
-37	24	98	109
-38	23	100	110
-39	23	102	111
-40	22	104	112
-41	21	106	113
-42	21	108	114
-43	20	110	115
-44	19	112	115
-45	18	113	116
-46	18	115	117
-47	17	117	118
-48	17	119	119
-49	16	121	119
-50	16	123	120
-51	16	125	121
-52	17	127	122
-53	17	128	122
-54	18	130	123
-55	20	132	123
-56	21	134	124
-57	23	136	124
-58	26	138	125
-59	28	139	125
-60	31	141	126
-61	34	143	126
-62	37	145	126
-63	41	146	127
-64	44	148	127
-65	48	150	127
-66	52	151	128
-67	56	153	128
-68	60	154	129
-69	64	156	129
-70	68	157	130
-71	72	159	130
-72	76	160	131
-73	80	162	131
-74	84	163	132
-75	88	165	133
-76	92	166	133
-77	95	168	134
-78	99	169	135
-79	103	170	136
-80	107	172	137
-81	111	173	138
-82	115	174	140
-83	118	176	141
-84	122	177	142
-85	125	178	143
-86	129	180	145
-87	133	181	146
-88	136	183	148
-89	140	184	150
-90	143	185	151
-91	146	187	153
-92	150	188	155
-93	153	189	156
-94	156	191	158
-95	160	192	160
-96	163	194	162
-97	166	195	164
-98	169	196	166
-99	173	198	169
-100	176	199	171
-101	179	201	173
-102	182	202	175
-103	185	204	177
-104	188	205	180
-105	191	207	182
-106	194	208	185
-107	197	210	187
-108	200	212	190
-109	203	213	192
-110	206	215	195
-111	209	216	197
-112	212	218	200
-113	214	220	203
-114	217	221	205
-115	220	223	208
-116	223	225	211
-117	226	226	214
-118	229	228	216
-119	231	230	219
-120	234	232	222
-121	237	233	225
-122	240	235	228
-123	243	237	231
-124	245	239	234
-125	248	241	237
-126	251	243	240
-127	253	245	243
-128	253	245	243
-129	251	242	240
-130	250	240	237
-131	249	237	233
-132	248	235	230
-133	247	233	227
-134	246	230	223
-135	245	228	220
-136	244	225	216
-137	243	223	213
-138	242	220	210
-139	241	218	206
-140	240	215	203
-141	239	213	200
-142	238	210	196
-143	237	208	193
-144	236	205	190
-145	236	203	187
-146	235	200	183
-147	234	198	180
-148	233	195	177
-149	233	193	174
-150	232	190	171
-151	231	188	168
-152	231	186	165
-153	230	183	162
-154	229	181	159
-155	229	178	156
-156	228	176	153
-157	227	173	150
-158	227	171	147
-159	226	168	145
-160	225	166	142
-161	225	163	139
-162	224	160	137
-163	223	158	134
-164	223	155	132
-165	222	153	129
-166	221	150	127
-167	221	148	125
-168	220	145	123
-169	219	143	121
-170	219	140	119
-171	218	138	117
-172	217	135	115
-173	216	133	113
-174	215	130	112
-175	215	128	110
-176	214	125	109
-177	213	123	107
-178	212	120	106
-179	211	118	105
-180	210	115	104
-181	209	113	103
-182	208	111	102
-183	206	108	101
-184	205	106	100
-185	204	104	100
-186	203	101	99
-187	201	99	98
-188	200	97	98
-189	199	94	97
-190	197	92	97
-191	196	90	97
-192	194	88	96
-193	193	86	96
-194	191	83	96
-195	189	81	96
-196	188	79	96
-197	186	77	96
-198	184	75	96
-199	183	73	95
-200	181	71	95
-201	179	69	95
-202	177	67	95
-203	176	65	95
-204	174	63	95
-205	172	61	96
-206	170	60	96
-207	168	58	96
-208	166	56	96
-209	164	54	96
-210	162	52	96
-211	160	51	96
-212	158	49	96
-213	156	47	96
-214	154	45	96
-215	151	44	96
-216	149	42	96
-217	147	41	96
-218	145	39	96
-219	143	38	96
-220	140	36	96
-221	138	35	96
-222	136	33	96
-223	133	32	95
-224	131	31	95
-225	129	30	95
-226	126	28	95
-227	124	27	94
-228	121	26	94
-229	119	25	93
-230	116	25	93
-231	114	24	92
-232	111	23	91
-233	109	22	90
-234	106	22	89
-235	103	21	88
-236	101	21	87
-237	98	20	86
-238	95	20	85
-239	93	20	83
-240	90	19	82
-241	87	19	80
-242	85	19	79
-243	82	18	77
-244	79	18	75
-245	77	18	74
-246	74	17	72
-247	72	17	70
-248	69	16	68
-249	66	16	66
-250	64	16	64
-251	61	15	61
-252	59	14	59
-253	56	14	57
-254	54	13	55
-255	51	13	53
