--- src/main/resources/fiji/plugin/trackmate/gui/displaysettings/luts/Deep.lut
@@ -1,257 +1,0 @@
-Index	Red	Green	Blue
-0	39	26	44
-1	40	26	45
-2	41	27	46
-3	41	28	48
-4	42	29	49
-5	43	29	50
-6	43	30	52
-7	44	31	53
-8	45	32	55
-9	45	32	56
-10	46	33	57
-11	47	34	59
-12	47	35	60
-13	48	35	62
-14	49	36	63
-15	49	37	65
-16	50	38	66
-17	50	38	68
-18	51	39	69
-19	52	40	71
-20	52	41	72
-21	53	41	74
-22	53	42	75
-23	54	43	77
-24	55	43	78
-25	55	44	80
-26	56	45	82
-27	56	46	83
-28	57	46	85
-29	57	47	86
-30	58	48	88
-31	58	48	90
-32	59	49	91
-33	59	50	93
-34	60	51	95
-35	60	51	96
-36	61	52	98
-37	61	53	99
-38	61	54	101
-39	62	54	103
-40	62	55	105
-41	62	56	106
-42	63	57	108
-43	63	57	110
-44	63	58	111
-45	64	59	113
-46	64	60	115
-47	64	61	116
-48	64	61	118
-49	65	62	120
-50	65	63	121
-51	65	64	123
-52	65	65	124
-53	65	66	126
-54	65	67	127
-55	65	68	129
-56	65	69	130
-57	65	69	132
-58	65	70	133
-59	65	71	134
-60	64	72	135
-61	64	74	136
-62	64	75	137
-63	64	76	138
-64	64	77	139
-65	63	78	140
-66	63	79	141
-67	63	80	142
-68	63	81	142
-69	62	82	143
-70	62	83	143
-71	62	84	144
-72	62	85	144
-73	62	87	145
-74	62	88	145
-75	61	89	145
-76	61	90	146
-77	61	91	146
-78	61	92	146
-79	61	93	147
-80	61	94	147
-81	61	95	147
-82	61	96	147
-83	61	97	148
-84	61	98	148
-85	61	99	148
-86	61	100	148
-87	61	101	149
-88	61	102	149
-89	62	103	149
-90	62	105	149
-91	62	106	149
-92	62	107	150
-93	62	108	150
-94	62	109	150
-95	62	110	150
-96	63	111	150
-97	63	112	151
-98	63	113	151
-99	63	114	151
-100	64	115	151
-101	64	116	151
-102	64	117	152
-103	64	118	152
-104	64	119	152
-105	65	120	152
-106	65	121	153
-107	65	122	153
-108	66	123	153
-109	66	124	153
-110	66	125	153
-111	66	126	154
-112	67	127	154
-113	67	128	154
-114	67	129	154
-115	68	129	154
-116	68	130	155
-117	68	131	155
-118	69	132	155
-119	69	133	155
-120	69	134	156
-121	70	135	156
-122	70	136	156
-123	70	137	156
-124	71	138	157
-125	71	139	157
-126	71	140	157
-127	72	141	157
-128	72	142	157
-129	72	143	158
-130	73	144	158
-131	73	145	158
-132	73	146	158
-133	74	147	159
-134	74	148	159
-135	74	149	159
-136	75	150	159
-137	75	151	159
-138	75	152	160
-139	76	153	160
-140	76	154	160
-141	76	155	160
-142	77	156	160
-143	77	157	161
-144	78	159	161
-145	78	160	161
-146	78	161	161
-147	79	162	161
-148	79	163	162
-149	80	164	162
-150	80	165	162
-151	81	166	162
-152	81	167	162
-153	81	168	162
-154	82	169	162
-155	82	170	163
-156	83	171	163
-157	83	172	163
-158	84	173	163
-159	85	174	163
-160	85	175	163
-161	86	176	163
-162	86	177	163
-163	87	178	163
-164	88	179	163
-165	88	180	163
-166	89	181	163
-167	90	182	163
-168	91	183	163
-169	92	184	163
-170	92	185	163
-171	93	186	163
-172	94	187	163
-173	95	188	163
-174	96	189	163
-175	97	190	163
-176	98	191	163
-177	99	192	163
-178	100	193	163
-179	102	194	163
-180	103	195	163
-181	104	196	163
-182	105	197	163
-183	107	198	163
-184	108	199	163
-185	110	200	163
-186	111	201	163
-187	113	202	163
-188	114	202	163
-189	116	203	163
-190	118	204	163
-191	119	205	162
-192	121	206	162
-193	123	207	162
-194	125	208	162
-195	127	209	162
-196	129	209	162
-197	131	210	162
-198	133	211	162
-199	135	212	163
-200	137	213	163
-201	139	213	163
-202	141	214	163
-203	143	215	163
-204	145	216	163
-205	147	216	163
-206	149	217	164
-207	152	218	164
-208	154	219	164
-209	156	219	165
-210	158	220	165
-211	160	221	165
-212	163	222	166
-213	165	222	166
-214	167	223	167
-215	169	224	167
-216	171	225	168
-217	174	225	168
-218	176	226	169
-219	178	227	169
-220	180	227	170
-221	182	228	171
-222	184	229	171
-223	187	230	172
-224	189	230	173
-225	191	231	173
-226	193	232	174
-227	195	232	175
-228	197	233	176
-229	199	234	176
-230	201	235	177
-231	204	235	178
-232	206	236	179
-233	208	237	180
-234	210	237	181
-235	212	238	182
-236	214	239	183
-237	216	240	184
-238	218	240	185
-239	220	241	186
-240	222	242	187
-241	224	242	188
-242	226	243	189
-243	228	244	190
-244	230	245	191
-245	232	245	192
-246	235	246	193
-247	237	247	194
-248	239	248	195
-249	241	248	196
-250	243	249	198
-251	245	250	199
-252	247	251	200
-253	249	252	201
-254	251	252	202
-255	253	253	204
