--- src/main/resources/fiji/plugin/trackmate/gui/displaysettings/luts/Delta.lut
@@ -1,257 +1,0 @@
-Index	Red	Green	Blue
-0	16	31	63
-1	18	32	66
-2	19	34	69
-3	20	35	73
-4	22	37	76
-5	23	38	79
-6	24	39	82
-7	25	41	86
-8	27	42	89
-9	28	43	92
-10	29	45	96
-11	30	46	99
-12	31	47	103
-13	32	48	107
-14	33	50	110
-15	34	51	114
-16	35	52	118
-17	36	53	122
-18	37	55	125
-19	37	56	129
-20	38	57	133
-21	38	59	137
-22	38	60	141
-23	38	62	144
-24	37	64	148
-25	35	66	150
-26	34	68	152
-27	33	70	153
-28	31	72	154
-29	30	75	155
-30	29	77	155
-31	29	79	156
-32	28	81	156
-33	27	83	157
-34	27	86	157
-35	27	88	157
-36	26	90	157
-37	26	92	158
-38	26	94	158
-39	26	96	158
-40	27	98	159
-41	27	100	159
-42	27	102	159
-43	28	104	160
-44	28	106	160
-45	29	108	160
-46	30	110	161
-47	31	112	161
-48	31	114	161
-49	32	116	162
-50	33	118	162
-51	34	120	163
-52	35	121	163
-53	36	123	164
-54	38	125	164
-55	39	127	165
-56	40	129	165
-57	41	131	165
-58	42	133	166
-59	44	135	166
-60	45	137	167
-61	46	139	167
-62	48	141	168
-63	49	142	168
-64	51	144	169
-65	53	146	169
-66	54	148	170
-67	56	150	170
-68	58	152	171
-69	60	154	171
-70	62	156	172
-71	64	157	172
-72	66	159	173
-73	68	161	173
-74	71	163	174
-75	73	165	174
-76	76	167	174
-77	79	168	175
-78	82	170	175
-79	86	172	176
-80	89	173	176
-81	93	175	176
-82	96	177	177
-83	100	178	177
-84	104	180	178
-85	108	181	179
-86	112	183	179
-87	116	184	180
-88	120	185	181
-89	124	187	182
-90	128	188	183
-91	132	190	184
-92	136	191	185
-93	140	193	186
-94	143	194	187
-95	147	196	188
-96	151	197	190
-97	154	198	191
-98	158	200	192
-99	162	201	193
-100	165	203	195
-101	169	204	196
-102	172	206	197
-103	175	208	199
-104	179	209	200
-105	182	211	202
-106	185	212	203
-107	189	214	204
-108	192	215	206
-109	195	217	207
-110	199	219	209
-111	202	220	210
-112	205	222	212
-113	208	224	213
-114	211	226	215
-115	214	227	216
-116	217	229	218
-117	221	231	219
-118	224	233	220
-119	227	235	222
-120	230	237	223
-121	233	238	224
-122	236	240	225
-123	239	242	226
-124	243	244	227
-125	246	246	228
-126	249	248	229
-127	253	250	229
-128	254	252	203
-129	253	249	199
-130	251	247	195
-131	250	244	191
-132	249	242	187
-133	247	240	182
-134	246	237	178
-135	245	235	174
-136	243	233	169
-137	242	230	165
-138	241	228	161
-139	239	226	156
-140	238	223	152
-141	236	221	148
-142	235	219	143
-143	233	217	139
-144	232	215	135
-145	230	213	130
-146	229	210	126
-147	227	208	122
-148	226	206	117
-149	224	204	113
-150	222	202	109
-151	220	201	104
-152	218	199	100
-153	216	197	96
-154	214	195	91
-155	212	193	87
-156	210	192	83
-157	207	190	79
-158	205	188	75
-159	203	187	71
-160	200	185	67
-161	197	184	63
-162	195	182	59
-163	192	181	56
-164	189	179	52
-165	186	178	48
-166	183	177	45
-167	180	175	42
-168	177	174	39
-169	174	173	35
-170	170	172	32
-171	167	170	29
-172	164	169	27
-173	160	168	24
-174	157	167	21
-175	154	166	19
-176	150	164	16
-177	147	163	14
-178	143	162	12
-179	140	161	10
-180	136	160	8
-181	133	158	7
-182	129	157	6
-183	126	156	5
-184	122	155	5
-185	119	153	5
-186	115	152	5
-187	112	151	6
-188	108	150	7
-189	104	149	8
-190	101	147	9
-191	97	146	11
-192	93	145	12
-193	90	143	14
-194	86	142	15
-195	82	141	17
-196	79	139	19
-197	75	138	20
-198	71	137	22
-199	68	135	23
-200	64	134	25
-201	61	132	26
-202	57	131	28
-203	54	129	29
-204	50	128	30
-205	47	126	32
-206	43	125	33
-207	40	123	34
-208	37	122	35
-209	34	120	36
-210	31	118	37
-211	28	117	38
-212	25	115	39
-213	23	113	40
-214	20	112	40
-215	18	110	41
-216	16	108	42
-217	14	106	42
-218	13	105	43
-219	12	103	43
-220	11	101	44
-221	10	99	44
-222	10	97	44
-223	11	95	44
-224	11	94	44
-225	12	92	44
-226	12	90	44
-227	13	88	44
-228	14	86	44
-229	15	84	44
-230	16	82	43
-231	17	80	43
-232	18	78	43
-233	19	76	42
-234	20	75	42
-235	20	73	41
-236	21	71	40
-237	22	69	40
-238	22	67	39
-239	23	65	38
-240	23	63	37
-241	24	61	36
-242	24	59	35
-243	24	57	34
-244	25	55	33
-245	25	53	32
-246	25	52	31
-247	25	50	30
-248	25	48	28
-249	25	46	27
-250	24	44	26
-251	24	42	24
-252	24	40	23
-253	23	38	21
-254	23	36	20
-255	23	35	18
