--- src/main/resources/fiji/plugin/trackmate/gui/displaysettings/luts/Dense.lut
@@ -1,257 +1,0 @@
-Index	Red	Green	Blue
-0	54	14	36
-1	55	14	37
-2	56	14	38
-3	57	15	39
-4	59	15	40
-5	60	15	42
-6	61	15	43
-7	62	16	44
-8	64	16	45
-9	65	16	47
-10	66	16	48
-11	67	17	49
-12	68	17	50
-13	69	17	52
-14	71	17	53
-15	72	18	55
-16	73	18	56
-17	74	18	57
-18	75	18	59
-19	76	19	60
-20	77	19	62
-21	78	19	63
-22	79	19	65
-23	80	20	66
-24	81	20	68
-25	82	20	69
-26	83	21	71
-27	84	21	73
-28	85	22	74
-29	86	22	76
-30	87	22	77
-31	88	23	79
-32	89	23	81
-33	90	24	82
-34	91	24	84
-35	92	25	86
-36	92	25	87
-37	93	26	89
-38	94	26	91
-39	95	27	92
-40	96	28	94
-41	96	28	96
-42	97	29	97
-43	98	29	99
-44	99	30	101
-45	99	31	102
-46	100	31	104
-47	101	32	106
-48	101	33	107
-49	102	34	109
-50	103	34	110
-51	103	35	112
-52	104	36	114
-53	105	37	115
-54	105	37	117
-55	106	38	119
-56	106	39	120
-57	107	40	122
-58	107	41	123
-59	108	42	125
-60	108	42	127
-61	109	43	128
-62	109	44	130
-63	110	45	131
-64	110	46	133
-65	111	47	134
-66	111	48	136
-67	112	49	138
-68	112	49	139
-69	113	50	141
-70	113	51	142
-71	113	52	144
-72	114	53	145
-73	114	54	147
-74	115	55	148
-75	115	56	150
-76	115	57	151
-77	116	58	153
-78	116	59	154
-79	116	60	155
-80	116	61	157
-81	117	62	158
-82	117	63	160
-83	117	64	161
-84	118	65	163
-85	118	66	164
-86	118	67	165
-87	118	68	167
-88	119	69	168
-89	119	70	170
-90	119	71	171
-91	119	72	172
-92	119	73	174
-93	119	74	175
-94	120	75	176
-95	120	76	178
-96	120	77	179
-97	120	78	180
-98	120	80	181
-99	120	81	183
-100	120	82	184
-101	120	83	185
-102	120	84	186
-103	121	85	188
-104	121	86	189
-105	121	87	190
-106	121	88	191
-107	121	89	192
-108	121	91	194
-109	121	92	195
-110	121	93	196
-111	121	94	197
-112	121	95	198
-113	121	96	199
-114	121	97	200
-115	120	99	201
-116	120	100	202
-117	120	101	203
-118	120	102	204
-119	120	103	205
-120	120	104	206
-121	120	105	207
-122	120	107	208
-123	120	108	209
-124	119	109	210
-125	119	110	211
-126	119	111	212
-127	119	113	213
-128	119	114	213
-129	119	115	214
-130	118	116	215
-131	118	117	216
-132	118	119	216
-133	118	120	217
-134	118	121	218
-135	118	122	218
-136	117	123	219
-137	117	125	220
-138	117	126	220
-139	117	127	221
-140	117	128	221
-141	116	129	222
-142	116	131	222
-143	116	132	223
-144	116	133	223
-145	116	134	224
-146	116	135	224
-147	115	137	225
-148	115	138	225
-149	115	139	225
-150	115	140	226
-151	115	141	226
-152	115	143	226
-153	115	144	227
-154	115	145	227
-155	115	146	227
-156	115	147	227
-157	115	148	227
-158	115	150	227
-159	115	151	228
-160	115	152	228
-161	115	153	228
-162	115	154	228
-163	116	155	228
-164	116	156	228
-165	116	158	228
-166	116	159	228
-167	117	160	228
-168	117	161	228
-169	117	162	228
-170	118	163	228
-171	118	164	228
-172	119	165	228
-173	119	166	228
-174	120	167	228
-175	120	169	228
-176	121	170	228
-177	122	171	228
-178	122	172	228
-179	123	173	227
-180	124	174	227
-181	124	175	227
-182	125	176	227
-183	126	177	227
-184	127	178	227
-185	128	179	227
-186	129	180	227
-187	130	181	227
-188	131	182	227
-189	132	183	227
-190	133	184	226
-191	134	185	226
-192	135	186	226
-193	136	187	226
-194	137	188	226
-195	138	189	226
-196	139	190	226
-197	140	190	226
-198	141	191	226
-199	143	192	226
-200	144	193	226
-201	145	194	226
-202	146	195	226
-203	148	196	226
-204	149	197	226
-205	150	198	226
-206	152	199	226
-207	153	200	226
-208	154	201	226
-209	156	201	226
-210	157	202	226
-211	159	203	226
-212	160	204	226
-213	161	205	226
-214	163	206	226
-215	164	207	226
-216	166	208	226
-217	167	209	226
-218	169	209	226
-219	170	210	226
-220	172	211	227
-221	173	212	227
-222	175	213	227
-223	176	214	227
-224	178	215	227
-225	179	215	227
-226	181	216	228
-227	183	217	228
-228	184	218	228
-229	186	219	228
-230	187	220	228
-231	189	220	229
-232	191	221	229
-233	192	222	229
-234	194	223	230
-235	196	224	230
-236	197	225	230
-237	199	225	231
-238	201	226	231
-239	202	227	231
-240	204	228	232
-241	206	229	232
-242	207	230	233
-243	209	230	233
-244	211	231	234
-245	212	232	234
-246	214	233	235
-247	216	234	235
-248	217	235	236
-249	219	235	236
-250	221	236	237
-251	223	237	237
-252	224	238	238
-253	226	239	239
-254	228	239	239
-255	230	240	240
