--- src/main/resources/fiji/plugin/trackmate/gui/displaysettings/luts/Gray.lut
@@ -1,257 +1,0 @@
-Index	Red	Green	Blue
-0	0	0	0
-1	0	0	0
-2	0	0	0
-3	0	0	0
-4	0	0	0
-5	1	1	1
-6	1	1	1
-7	2	1	2
-8	2	2	2
-9	3	3	3
-10	3	3	3
-11	4	4	4
-12	5	5	5
-13	6	6	6
-14	7	7	7
-15	8	7	7
-16	9	8	8
-17	10	10	9
-18	11	11	11
-19	12	12	12
-20	13	13	13
-21	14	14	14
-22	15	15	15
-23	16	16	16
-24	17	17	17
-25	18	18	18
-26	19	19	19
-27	20	20	20
-28	21	21	21
-29	22	22	22
-30	23	23	23
-31	24	24	24
-32	25	25	25
-33	26	26	26
-34	27	27	27
-35	28	28	28
-36	29	29	28
-37	30	30	29
-38	31	30	30
-39	32	31	31
-40	33	32	32
-41	34	33	33
-42	35	34	34
-43	36	35	35
-44	37	36	36
-45	37	37	37
-46	38	38	38
-47	39	39	39
-48	40	40	40
-49	41	41	41
-50	42	42	41
-51	43	43	42
-52	44	43	43
-53	45	44	44
-54	46	45	45
-55	47	46	46
-56	48	47	47
-57	48	48	48
-58	49	49	49
-59	50	50	50
-60	51	51	51
-61	52	52	51
-62	53	52	52
-63	54	53	53
-64	55	54	54
-65	56	55	55
-66	57	56	56
-67	57	57	57
-68	58	58	58
-69	59	59	59
-70	60	60	59
-71	61	61	60
-72	62	61	61
-73	63	62	62
-74	64	63	63
-75	65	64	64
-76	66	65	65
-77	66	66	66
-78	67	67	67
-79	68	68	67
-80	69	69	68
-81	70	70	69
-82	71	70	70
-83	72	71	71
-84	73	72	72
-85	74	73	73
-86	75	74	74
-87	75	75	75
-88	76	76	76
-89	77	77	76
-90	78	78	77
-91	79	79	78
-92	80	79	79
-93	81	80	80
-94	82	81	81
-95	83	82	82
-96	84	83	83
-97	84	84	84
-98	85	85	84
-99	86	86	85
-100	87	87	86
-101	88	88	87
-102	89	89	88
-103	90	89	89
-104	91	90	90
-105	92	91	91
-106	93	92	92
-107	94	93	93
-108	94	94	94
-109	95	95	94
-110	96	96	95
-111	97	97	96
-112	98	98	97
-113	99	99	98
-114	100	100	99
-115	101	100	100
-116	102	101	101
-117	103	102	102
-118	104	103	103
-119	105	104	104
-120	106	105	105
-121	106	106	106
-122	107	107	106
-123	108	108	107
-124	109	109	108
-125	110	110	109
-126	111	111	110
-127	112	112	111
-128	113	113	112
-129	114	114	113
-130	115	114	114
-131	116	115	115
-132	117	116	116
-133	118	117	117
-134	119	118	118
-135	120	119	119
-136	121	120	120
-137	122	121	121
-138	123	122	122
-139	124	123	122
-140	125	124	123
-141	125	125	124
-142	126	126	125
-143	127	127	126
-144	128	128	127
-145	129	129	128
-146	130	130	129
-147	131	131	130
-148	132	132	131
-149	133	133	132
-150	134	134	133
-151	135	135	134
-152	136	136	135
-153	137	137	136
-154	138	138	137
-155	139	139	138
-156	140	140	139
-157	141	141	140
-158	142	142	141
-159	143	143	142
-160	144	144	143
-161	145	145	144
-162	146	146	145
-163	147	147	146
-164	148	148	147
-165	149	149	148
-166	150	150	149
-167	151	151	150
-168	152	152	151
-169	153	153	152
-170	154	154	153
-171	155	155	154
-172	156	156	155
-173	157	157	156
-174	158	158	157
-175	160	159	158
-176	161	160	159
-177	162	161	160
-178	163	162	161
-179	164	163	163
-180	165	164	164
-181	166	165	165
-182	167	167	166
-183	168	168	167
-184	169	169	168
-185	170	170	169
-186	171	171	170
-187	172	172	171
-188	173	173	172
-189	174	174	173
-190	176	175	174
-191	177	176	175
-192	178	177	176
-193	179	178	178
-194	180	180	179
-195	181	181	180
-196	182	182	181
-197	183	183	182
-198	184	184	183
-199	185	185	184
-200	187	186	185
-201	188	187	186
-202	189	189	188
-203	190	190	189
-204	191	191	190
-205	192	192	191
-206	193	193	192
-207	194	194	193
-208	196	195	194
-209	197	197	195
-210	198	198	197
-211	199	199	198
-212	200	200	199
-213	201	201	200
-214	203	202	201
-215	204	204	202
-216	205	205	204
-217	206	206	205
-218	207	207	206
-219	209	208	207
-220	210	209	208
-221	211	211	210
-222	212	212	211
-223	213	213	212
-224	215	214	213
-225	216	216	214
-226	217	217	216
-227	218	218	217
-228	219	219	218
-229	221	220	219
-230	222	222	220
-231	223	223	222
-232	224	224	223
-233	226	225	224
-234	227	227	225
-235	228	228	227
-236	229	229	228
-237	231	230	229
-238	232	232	230
-239	233	233	232
-240	235	234	233
-241	236	236	234
-242	237	237	236
-243	238	238	237
-244	240	239	238
-245	241	241	240
-246	242	242	241
-247	244	243	242
-248	245	245	243
-249	246	246	245
-250	248	247	246
-251	249	249	247
-252	250	250	249
-253	252	251	250
-254	253	253	251
-255	254	254	253
