--- src/main/resources/fiji/plugin/trackmate/gui/displaysettings/luts/Haline.lut
@@ -1,257 +1,0 @@
-Index	Red	Green	Blue
-0	41	24	107
-1	42	24	110
-2	42	24	112
-3	42	25	114
-4	43	25	117
-5	43	25	119
-6	44	25	122
-7	44	26	124
-8	44	26	127
-9	45	26	129
-10	45	26	132
-11	45	27	134
-12	45	27	137
-13	46	27	139
-14	46	28	142
-15	46	28	144
-16	46	29	147
-17	46	29	149
-18	45	30	151
-19	45	31	153
-20	45	32	155
-21	44	33	157
-22	43	34	159
-23	42	35	160
-24	41	37	161
-25	40	39	162
-26	39	40	162
-27	38	42	162
-28	37	44	162
-29	35	46	162
-30	34	48	162
-31	33	49	162
-32	31	51	161
-33	30	53	161
-34	29	55	160
-35	27	56	160
-36	26	58	159
-37	25	59	159
-38	24	61	158
-39	22	62	157
-40	21	64	157
-41	20	65	156
-42	19	66	155
-43	18	68	155
-44	17	69	154
-45	16	70	153
-46	15	71	153
-47	15	73	152
-48	14	74	152
-49	13	75	151
-50	13	76	150
-51	12	77	150
-52	12	78	149
-53	12	79	149
-54	12	81	148
-55	12	82	148
-56	12	83	147
-57	12	84	147
-58	12	85	146
-59	12	86	146
-60	13	87	145
-61	13	88	145
-62	14	89	144
-63	14	90	144
-64	15	91	144
-65	15	92	143
-66	16	93	143
-67	17	93	143
-68	18	94	142
-69	18	95	142
-70	19	96	142
-71	20	97	141
-72	21	98	141
-73	21	99	141
-74	22	100	140
-75	23	101	140
-76	24	102	140
-77	25	102	140
-78	26	103	139
-79	26	104	139
-80	27	105	139
-81	28	106	139
-82	29	107	139
-83	30	108	139
-84	31	109	138
-85	31	109	138
-86	32	110	138
-87	33	111	138
-88	34	112	138
-89	35	113	138
-90	36	114	137
-91	36	115	137
-92	37	115	137
-93	38	116	137
-94	39	117	137
-95	39	118	137
-96	40	119	137
-97	41	120	137
-98	42	121	137
-99	42	121	137
-100	43	122	137
-101	44	123	136
-102	44	124	136
-103	45	125	136
-104	46	126	136
-105	46	127	136
-106	47	127	136
-107	48	128	136
-108	48	129	136
-109	49	130	136
-110	49	131	136
-111	50	132	136
-112	51	133	136
-113	51	133	136
-114	52	134	136
-115	52	135	136
-116	53	136	136
-117	54	137	136
-118	54	138	135
-119	55	139	135
-120	55	140	135
-121	56	140	135
-122	56	141	135
-123	57	142	135
-124	57	143	135
-125	58	144	135
-126	58	145	135
-127	59	146	135
-128	59	147	135
-129	60	148	134
-130	60	148	134
-131	61	149	134
-132	61	150	134
-133	62	151	134
-134	62	152	134
-135	63	153	134
-136	64	154	133
-137	64	155	133
-138	65	156	133
-139	65	157	133
-140	66	158	132
-141	66	158	132
-142	67	159	132
-143	67	160	132
-144	68	161	131
-145	69	162	131
-146	69	163	131
-147	70	164	131
-148	70	165	130
-149	71	166	130
-150	72	167	130
-151	72	168	129
-152	73	169	129
-153	74	169	128
-154	74	170	128
-155	75	171	128
-156	76	172	127
-157	77	173	127
-158	78	174	126
-159	78	175	126
-160	79	176	125
-161	80	177	125
-162	81	178	124
-163	82	179	124
-164	83	179	123
-165	84	180	123
-166	85	181	122
-167	86	182	121
-168	87	183	121
-169	88	184	120
-170	89	185	120
-171	90	186	119
-172	91	187	118
-173	92	187	118
-174	94	188	117
-175	95	189	116
-176	96	190	115
-177	98	191	115
-178	99	192	114
-179	100	193	113
-180	102	193	112
-181	103	194	112
-182	105	195	111
-183	106	196	110
-184	108	197	109
-185	109	198	108
-186	111	198	107
-187	113	199	107
-188	114	200	106
-189	116	201	105
-190	118	201	104
-191	120	202	103
-192	122	203	102
-193	124	204	101
-194	126	204	100
-195	128	205	100
-196	130	206	99
-197	132	206	98
-198	134	207	97
-199	136	208	96
-200	138	208	96
-201	141	209	95
-202	143	210	94
-203	145	210	93
-204	148	211	93
-205	150	211	92
-206	152	212	92
-207	155	213	92
-208	157	213	91
-209	160	214	91
-210	162	214	91
-211	165	215	91
-212	167	215	91
-213	170	216	92
-214	172	216	92
-215	175	216	92
-216	177	217	93
-217	179	217	94
-218	182	218	94
-219	184	218	95
-220	186	219	96
-221	189	219	97
-222	191	220	98
-223	193	220	99
-224	195	221	100
-225	197	221	102
-226	199	222	103
-227	202	222	104
-228	204	223	106
-229	206	223	107
-230	208	224	109
-231	210	224	110
-232	212	225	112
-233	214	225	113
-234	216	226	115
-235	217	226	117
-236	219	227	118
-237	221	228	120
-238	223	228	122
-239	225	229	123
-240	227	229	125
-241	229	230	127
-242	230	230	129
-243	232	231	131
-244	234	232	132
-245	236	232	134
-246	238	233	136
-247	239	233	138
-248	241	234	140
-249	243	235	142
-250	245	235	144
-251	246	236	145
-252	248	236	147
-253	250	237	149
-254	251	238	151
-255	253	238	153
