--- src/main/resources/fiji/plugin/trackmate/gui/displaysettings/luts/Ice.lut
@@ -1,257 +1,0 @@
-Index	Red	Green	Blue
-0	3	5	18
-1	4	6	19
-2	5	7	21
-3	6	8	22
-4	6	8	24
-5	7	9	25
-6	8	10	27
-7	9	11	28
-8	10	12	30
-9	11	13	31
-10	12	14	32
-11	13	15	34
-12	14	16	35
-13	15	16	37
-14	16	17	38
-15	17	18	40
-16	18	19	41
-17	19	20	42
-18	20	21	44
-19	21	21	45
-20	22	22	47
-21	23	23	48
-22	24	24	50
-23	25	25	51
-24	26	25	53
-25	27	26	54
-26	27	27	56
-27	28	28	57
-28	29	29	59
-29	30	29	60
-30	31	30	62
-31	32	31	63
-32	33	32	65
-33	34	32	66
-34	34	33	68
-35	35	34	69
-36	36	35	71
-37	37	35	72
-38	38	36	74
-39	39	37	75
-40	39	38	77
-41	40	38	79
-42	41	39	80
-43	42	40	82
-44	42	41	83
-45	43	41	85
-46	44	42	87
-47	45	43	88
-48	45	44	90
-49	46	45	91
-50	47	45	93
-51	48	46	95
-52	48	47	96
-53	49	48	98
-54	50	48	100
-55	50	49	101
-56	51	50	103
-57	52	51	105
-58	52	51	106
-59	53	52	108
-60	53	53	110
-61	54	54	111
-62	55	55	113
-63	55	55	115
-64	56	56	116
-65	56	57	118
-66	57	58	120
-67	57	59	121
-68	58	59	123
-69	58	60	125
-70	58	61	126
-71	59	62	128
-72	59	63	130
-73	60	64	131
-74	60	65	133
-75	60	65	134
-76	60	66	136
-77	61	67	138
-78	61	68	139
-79	61	69	141
-80	61	70	142
-81	62	71	144
-82	62	72	145
-83	62	73	147
-84	62	74	148
-85	62	75	150
-86	62	76	151
-87	62	77	152
-88	62	78	154
-89	62	79	155
-90	62	80	156
-91	62	81	157
-92	62	82	159
-93	62	83	160
-94	62	84	161
-95	62	85	162
-96	62	87	163
-97	62	88	164
-98	62	89	165
-99	62	90	166
-100	62	91	167
-101	62	92	168
-102	62	93	168
-103	62	94	169
-104	62	96	170
-105	62	97	171
-106	62	98	172
-107	62	99	172
-108	62	100	173
-109	62	101	174
-110	62	102	174
-111	62	103	175
-112	62	105	175
-113	62	106	176
-114	62	107	177
-115	62	108	177
-116	62	109	178
-117	63	110	178
-118	63	111	179
-119	63	112	179
-120	63	114	180
-121	63	115	180
-122	64	116	180
-123	64	117	181
-124	64	118	181
-125	65	119	182
-126	65	120	182
-127	65	121	183
-128	66	122	183
-129	66	123	183
-130	67	125	184
-131	67	126	184
-132	68	127	185
-133	68	128	185
-134	69	129	185
-135	69	130	186
-136	70	131	186
-137	70	132	186
-138	71	133	187
-139	72	134	187
-140	72	135	188
-141	73	136	188
-142	74	137	188
-143	74	139	189
-144	75	140	189
-145	76	141	189
-146	76	142	190
-147	77	143	190
-148	78	144	190
-149	79	145	191
-150	79	146	191
-151	80	147	192
-152	81	148	192
-153	82	149	192
-154	82	150	193
-155	83	151	193
-156	84	152	193
-157	85	153	194
-158	86	154	194
-159	87	156	195
-160	88	157	195
-161	88	158	195
-162	89	159	196
-163	90	160	196
-164	91	161	197
-165	92	162	197
-166	93	163	197
-167	94	164	198
-168	95	165	198
-169	96	166	199
-170	97	167	199
-171	98	168	199
-172	99	169	200
-173	100	170	200
-174	101	171	201
-175	102	172	201
-176	103	174	201
-177	104	175	202
-178	105	176	202
-179	106	177	203
-180	107	178	203
-181	109	179	203
-182	110	180	204
-183	111	181	204
-184	112	182	205
-185	113	183	205
-186	114	184	205
-187	116	185	206
-188	117	186	206
-189	118	187	207
-190	119	188	207
-191	121	189	208
-192	122	190	208
-193	123	191	208
-194	125	192	209
-195	126	193	209
-196	128	194	210
-197	129	195	210
-198	131	196	211
-199	132	198	211
-200	134	199	211
-201	135	200	212
-202	137	201	212
-203	138	202	213
-204	140	203	213
-205	142	203	214
-206	143	204	214
-207	145	205	215
-208	147	206	215
-209	149	207	216
-210	150	208	216
-211	152	209	217
-212	154	210	217
-213	156	211	218
-214	158	212	219
-215	160	213	219
-216	161	214	220
-217	163	215	220
-218	165	216	221
-219	167	217	222
-220	169	218	222
-221	171	219	223
-222	173	220	224
-223	175	221	225
-224	176	221	225
-225	178	222	226
-226	180	223	227
-227	182	224	228
-228	184	225	228
-229	186	226	229
-230	188	227	230
-231	190	228	231
-232	192	229	232
-233	193	230	233
-234	195	231	234
-235	197	232	234
-236	199	233	235
-237	201	234	236
-238	203	235	237
-239	205	236	238
-240	207	237	239
-241	208	238	240
-242	210	239	241
-243	212	240	242
-244	214	241	243
-245	216	242	244
-246	218	243	245
-247	219	244	245
-248	221	245	246
-249	223	246	247
-250	225	247	248
-251	227	248	249
-252	228	249	250
-253	230	250	251
-254	232	251	252
-255	234	252	253
