--- src/main/resources/fiji/plugin/trackmate/gui/displaysettings/luts/Matter.lut
@@ -1,257 +1,0 @@
-Index	Red	Green	Blue
-0	47	15	61
-1	48	15	62
-2	49	15	63
-3	50	16	64
-4	52	16	65
-5	53	16	66
-6	54	17	67
-7	56	17	68
-8	57	17	68
-9	58	18	69
-10	59	18	70
-11	61	18	71
-12	62	18	72
-13	63	19	73
-14	64	19	73
-15	66	19	74
-16	67	19	75
-17	68	20	76
-18	70	20	76
-19	71	20	77
-20	72	20	78
-21	74	21	79
-22	75	21	79
-23	76	21	80
-24	78	21	81
-25	79	21	82
-26	80	22	82
-27	82	22	83
-28	83	22	84
-29	84	22	84
-30	85	22	85
-31	87	22	86
-32	88	23	86
-33	89	23	87
-34	91	23	87
-35	92	23	88
-36	94	23	89
-37	95	23	89
-38	96	23	90
-39	98	23	90
-40	99	24	91
-41	100	24	91
-42	102	24	92
-43	103	24	92
-44	104	24	93
-45	106	24	93
-46	107	24	93
-47	108	25	94
-48	110	25	94
-49	111	25	95
-50	112	25	95
-51	114	25	95
-52	115	25	96
-53	116	25	96
-54	118	26	96
-55	119	26	97
-56	120	26	97
-57	122	26	97
-58	123	26	97
-59	125	26	98
-60	126	27	98
-61	127	27	98
-62	129	27	98
-63	130	27	98
-64	131	28	98
-65	133	28	99
-66	134	28	99
-67	135	28	99
-68	136	29	99
-69	138	29	99
-70	139	29	99
-71	140	30	99
-72	142	30	99
-73	143	30	99
-74	144	31	99
-75	146	31	99
-76	147	31	99
-77	148	32	99
-78	149	32	99
-79	151	33	99
-80	152	33	99
-81	153	33	98
-82	155	34	98
-83	156	34	98
-84	157	35	98
-85	158	35	98
-86	160	36	98
-87	161	36	98
-88	162	37	97
-89	163	37	97
-90	165	38	97
-91	166	39	97
-92	167	39	97
-93	168	40	96
-94	169	40	96
-95	171	41	96
-96	172	42	96
-97	173	42	95
-98	174	43	95
-99	175	44	95
-100	177	44	95
-101	178	45	94
-102	179	46	94
-103	180	46	94
-104	181	47	93
-105	182	48	93
-106	184	48	93
-107	185	49	93
-108	186	50	92
-109	187	51	92
-110	188	51	92
-111	189	52	91
-112	190	53	91
-113	191	54	91
-114	192	55	90
-115	193	56	90
-116	195	56	90
-117	196	57	89
-118	197	58	89
-119	198	59	89
-120	199	60	88
-121	200	61	88
-122	201	62	88
-123	202	63	87
-124	203	64	87
-125	204	65	87
-126	205	66	86
-127	206	67	86
-128	207	68	86
-129	208	69	85
-130	208	70	85
-131	209	71	85
-132	210	72	84
-133	211	73	84
-134	212	74	84
-135	213	75	84
-136	214	76	83
-137	215	78	83
-138	215	79	83
-139	216	80	83
-140	217	81	83
-141	218	82	83
-142	218	83	82
-143	219	85	82
-144	220	86	82
-145	221	87	82
-146	221	88	82
-147	222	89	82
-148	223	91	82
-149	223	92	82
-150	224	93	82
-151	225	95	82
-152	225	96	82
-153	226	97	82
-154	226	98	82
-155	227	100	83
-156	227	101	83
-157	228	102	83
-158	229	104	83
-159	229	105	83
-160	230	106	84
-161	230	108	84
-162	231	109	84
-163	231	110	85
-164	232	112	85
-165	232	113	85
-166	232	114	86
-167	233	116	86
-168	233	117	87
-169	234	118	87
-170	234	120	88
-171	235	121	88
-172	235	123	89
-173	235	124	89
-174	236	125	90
-175	236	127	91
-176	237	128	91
-177	237	129	92
-178	237	131	92
-179	238	132	93
-180	238	133	94
-181	238	135	94
-182	239	136	95
-183	239	138	96
-184	239	139	97
-185	240	140	97
-186	240	142	98
-187	240	143	99
-188	241	144	100
-189	241	146	101
-190	241	147	101
-191	241	148	102
-192	242	150	103
-193	242	151	104
-194	242	153	105
-195	243	154	106
-196	243	155	107
-197	243	157	108
-198	243	158	108
-199	244	159	109
-200	244	161	110
-201	244	162	111
-202	244	163	112
-203	245	165	113
-204	245	166	114
-205	245	168	115
-206	245	169	116
-207	245	170	117
-208	246	172	118
-209	246	173	119
-210	246	174	120
-211	246	176	121
-212	247	177	122
-213	247	178	124
-214	247	180	125
-215	247	181	126
-216	247	183	127
-217	247	184	128
-218	248	185	129
-219	248	187	130
-220	248	188	131
-221	248	189	132
-222	248	191	134
-223	249	192	135
-224	249	194	136
-225	249	195	137
-226	249	196	138
-227	249	198	139
-228	249	199	141
-229	250	200	142
-230	250	202	143
-231	250	203	144
-232	250	205	145
-233	250	206	147
-234	250	207	148
-235	250	209	149
-236	251	210	150
-237	251	212	152
-238	251	213	153
-239	251	214	154
-240	251	216	156
-241	251	217	157
-242	251	219	158
-243	252	220	159
-244	252	221	161
-245	252	223	162
-246	252	224	163
-247	252	226	165
-248	252	227	166
-249	252	228	167
-250	252	230	169
-251	253	231	170
-252	253	233	172
-253	253	234	173
-254	253	235	174
-255	253	237	176
