--- src/main/resources/fiji/plugin/trackmate/gui/displaysettings/luts/Oxy.lut
@@ -1,257 +1,0 @@
-Index	Red	Green	Blue
-0	63	5	5
-1	65	5	5
-2	67	5	5
-3	68	5	6
-4	70	6	6
-5	72	6	7
-6	73	6	7
-7	75	6	7
-8	77	6	8
-9	78	6	8
-10	80	6	9
-11	82	6	9
-12	83	7	10
-13	85	7	10
-14	87	7	11
-15	89	7	11
-16	90	7	11
-17	92	7	12
-18	94	7	12
-19	95	7	12
-20	97	7	13
-21	99	6	13
-22	101	6	13
-23	102	6	14
-24	104	6	14
-25	106	6	14
-26	107	6	14
-27	109	6	14
-28	111	5	15
-29	113	5	15
-30	114	5	15
-31	116	5	15
-32	118	5	15
-33	119	4	14
-34	121	4	14
-35	123	4	14
-36	124	5	14
-37	126	5	13
-38	128	6	13
-39	129	6	12
-40	131	7	12
-41	132	9	11
-42	133	10	10
-43	135	12	10
-44	136	14	9
-45	137	16	9
-46	138	17	8
-47	139	19	8
-48	141	21	8
-49	142	23	7
-50	143	24	7
-51	92	68	64
-52	80	79	79
-53	81	80	80
-54	82	81	81
-55	83	82	82
-56	84	83	83
-57	85	84	84
-58	86	85	85
-59	86	86	86
-60	87	87	87
-61	88	88	87
-62	89	89	88
-63	90	90	89
-64	91	91	90
-65	92	92	91
-66	93	92	92
-67	94	93	93
-68	95	94	94
-69	96	95	95
-70	97	96	96
-71	98	97	97
-72	99	98	98
-73	100	99	99
-74	101	100	100
-75	102	101	101
-76	102	102	102
-77	103	103	102
-78	104	104	103
-79	105	105	104
-80	106	106	105
-81	107	107	106
-82	108	108	107
-83	109	109	108
-84	110	110	109
-85	111	111	110
-86	112	112	111
-87	113	113	112
-88	114	114	113
-89	115	114	114
-90	116	115	115
-91	117	116	116
-92	118	117	117
-93	119	118	118
-94	120	119	119
-95	121	120	120
-96	122	121	121
-97	123	122	122
-98	124	123	123
-99	125	124	124
-100	126	125	125
-101	127	126	126
-102	128	127	127
-103	129	128	128
-104	130	129	129
-105	131	130	130
-106	132	131	131
-107	133	132	132
-108	134	133	133
-109	135	134	134
-110	136	135	135
-111	137	136	136
-112	138	137	137
-113	139	138	138
-114	140	139	139
-115	141	140	140
-116	142	141	141
-117	143	143	142
-118	144	144	143
-119	145	145	144
-120	146	146	145
-121	147	147	146
-122	148	148	147
-123	149	149	148
-124	150	150	149
-125	151	151	150
-126	152	152	151
-127	153	153	152
-128	154	154	153
-129	156	155	154
-130	157	156	155
-131	158	157	156
-132	159	158	157
-133	160	159	159
-134	161	160	160
-135	162	162	161
-136	163	163	162
-137	164	164	163
-138	165	165	164
-139	166	166	165
-140	167	167	166
-141	168	168	167
-142	170	169	168
-143	171	170	169
-144	172	171	171
-145	173	173	172
-146	174	174	173
-147	175	175	174
-148	176	176	175
-149	177	177	176
-150	178	178	177
-151	180	179	178
-152	181	180	179
-153	182	182	181
-154	183	183	182
-155	184	184	183
-156	185	185	184
-157	186	186	185
-158	188	187	186
-159	189	188	187
-160	190	190	189
-161	191	191	190
-162	192	192	191
-163	193	193	192
-164	195	194	193
-165	196	195	194
-166	197	197	196
-167	198	198	197
-168	199	199	198
-169	201	200	199
-170	202	201	200
-171	203	203	202
-172	204	204	203
-173	205	205	204
-174	207	206	205
-175	208	208	206
-176	209	209	208
-177	210	210	209
-178	211	211	210
-179	213	212	211
-180	214	214	213
-181	215	215	214
-182	216	216	215
-183	218	217	216
-184	219	219	218
-185	220	220	219
-186	222	221	220
-187	223	223	221
-188	224	224	223
-189	225	225	224
-190	227	226	225
-191	228	228	226
-192	229	229	228
-193	231	230	229
-194	232	232	230
-195	233	233	232
-196	234	234	233
-197	236	236	234
-198	237	237	236
-199	238	238	237
-200	240	240	238
-201	241	241	240
-202	242	242	241
-203	244	244	242
-204	245	247	216
-205	247	253	103
-206	245	252	100
-207	243	251	97
-208	242	250	94
-209	240	249	91
-210	238	247	88
-211	237	246	84
-212	236	245	80
-213	235	243	76
-214	234	242	73
-215	234	240	69
-216	234	239	66
-217	233	237	63
-218	233	235	61
-219	233	233	59
-220	233	232	57
-221	233	230	55
-222	233	228	53
-223	233	226	52
-224	232	225	51
-225	232	223	49
-226	232	221	48
-227	232	220	47
-228	231	218	46
-229	231	216	45
-230	231	214	44
-231	230	213	43
-232	230	211	42
-233	230	209	41
-234	229	208	40
-235	229	206	39
-236	229	205	38
-237	228	203	38
-238	228	201	37
-239	227	200	36
-240	227	198	35
-241	227	196	34
-242	226	195	34
-243	226	193	33
-244	225	192	32
-245	225	190	32
-246	224	188	31
-247	224	187	30
-248	223	185	30
-249	223	184	29
-250	223	182	28
-251	222	181	28
-252	222	179	27
-253	221	178	26
-254	221	176	26
-255	220	174	25
