--- src/main/resources/fiji/plugin/trackmate/gui/displaysettings/luts/Phase.lut
@@ -1,257 +1,0 @@
-Index	Red	Green	Blue
-0	167	119	12
-1	169	118	14
-2	170	117	16
-3	172	116	18
-4	173	115	20
-5	175	115	22
-6	176	114	23
-7	178	113	25
-8	179	112	27
-9	180	111	28
-10	182	110	30
-11	183	109	31
-12	184	107	33
-13	186	106	35
-14	187	105	36
-15	188	104	38
-16	189	103	39
-17	190	102	41
-18	192	101	43
-19	193	100	44
-20	194	99	46
-21	195	98	47
-22	196	97	49
-23	197	96	51
-24	198	95	52
-25	199	93	54
-26	200	92	56
-27	201	91	58
-28	202	90	59
-29	203	89	61
-30	204	88	63
-31	205	86	65
-32	206	85	67
-33	207	84	68
-34	208	83	70
-35	209	81	72
-36	210	80	74
-37	211	79	76
-38	211	78	78
-39	212	76	80
-40	213	75	83
-41	214	74	85
-42	214	72	87
-43	215	71	89
-44	216	69	91
-45	216	68	94
-46	217	67	96
-47	218	65	99
-48	218	64	101
-49	219	62	104
-50	219	61	106
-51	220	59	109
-52	220	58	112
-53	220	57	114
-54	221	55	117
-55	221	54	120
-56	221	52	123
-57	222	51	126
-58	222	49	129
-59	222	48	132
-60	222	47	135
-61	222	45	138
-62	222	44	141
-63	222	43	144
-64	222	42	147
-65	222	41	150
-66	222	40	153
-67	221	39	156
-68	221	38	160
-69	221	38	163
-70	220	37	166
-71	220	37	169
-72	219	37	172
-73	219	37	175
-74	218	37	178
-75	217	37	181
-76	216	38	184
-77	216	38	187
-78	215	39	190
-79	214	40	193
-80	213	41	195
-81	212	42	198
-82	211	43	201
-83	210	44	203
-84	209	45	206
-85	207	47	208
-86	206	48	210
-87	205	50	212
-88	204	51	215
-89	202	53	217
-90	201	54	219
-91	199	56	220
-92	198	58	222
-93	196	60	224
-94	195	61	226
-95	193	63	227
-96	192	65	229
-97	190	66	230
-98	188	68	232
-99	187	70	233
-100	185	71	234
-101	183	73	235
-102	181	75	236
-103	180	77	237
-104	178	78	238
-105	176	80	239
-106	174	82	240
-107	172	83	240
-108	170	85	241
-109	168	86	242
-110	166	88	242
-111	164	90	242
-112	162	91	243
-113	160	93	243
-114	158	94	243
-115	156	96	243
-116	153	97	244
-117	151	99	244
-118	149	100	244
-119	147	102	243
-120	144	103	243
-121	142	105	243
-122	140	106	243
-123	137	107	242
-124	135	109	242
-125	132	110	242
-126	130	112	241
-127	127	113	240
-128	125	114	240
-129	122	115	239
-130	120	117	238
-131	117	118	237
-132	115	119	236
-133	112	120	235
-134	109	122	234
-135	106	123	233
-136	104	124	232
-137	101	125	230
-138	98	126	229
-139	95	127	228
-140	93	128	226
-141	90	129	225
-142	87	130	223
-143	84	131	221
-144	82	132	220
-145	79	133	218
-146	76	134	216
-147	73	135	214
-148	71	136	212
-149	68	137	210
-150	66	137	208
-151	63	138	206
-152	61	139	204
-153	58	139	202
-154	56	140	200
-155	53	141	198
-156	51	141	196
-157	49	142	194
-158	47	142	192
-159	45	143	190
-160	43	143	188
-161	42	144	185
-162	40	144	183
-163	38	144	181
-164	37	145	179
-165	35	145	177
-166	34	146	175
-167	33	146	173
-168	32	146	171
-169	31	147	169
-170	30	147	167
-171	29	147	165
-172	28	147	163
-173	27	148	161
-174	26	148	159
-175	25	148	157
-176	24	148	156
-177	23	149	154
-178	22	149	152
-179	22	149	150
-180	21	149	148
-181	20	150	146
-182	19	150	144
-183	18	150	142
-184	17	150	140
-185	16	151	138
-186	15	151	136
-187	15	151	134
-188	14	151	132
-189	13	151	130
-190	12	152	128
-191	12	152	125
-192	11	152	123
-193	11	152	121
-194	11	152	119
-195	11	153	116
-196	11	153	114
-197	12	153	112
-198	13	153	109
-199	15	153	107
-200	16	153	104
-201	18	153	102
-202	20	153	99
-203	22	153	96
-204	25	154	93
-205	27	154	91
-206	30	154	88
-207	33	154	85
-208	36	153	82
-209	39	153	79
-210	42	153	76
-211	45	153	73
-212	49	153	69
-213	52	153	66
-214	56	152	63
-215	59	152	60
-216	63	152	56
-217	67	151	53
-218	71	151	50
-219	74	150	46
-220	78	150	43
-221	82	149	40
-222	86	149	37
-223	90	148	34
-224	93	147	32
-225	97	146	29
-226	100	146	27
-227	104	145	25
-228	107	144	23
-229	110	143	21
-230	113	142	19
-231	116	141	18
-232	119	141	17
-233	121	140	16
-234	124	139	15
-235	127	138	15
-236	129	137	14
-237	131	136	14
-238	134	135	13
-239	136	135	13
-240	138	134	13
-241	140	133	13
-242	142	132	13
-243	145	131	13
-244	147	130	13
-245	149	129	13
-246	151	128	13
-247	153	127	13
-248	154	126	13
-249	156	125	13
-250	158	124	13
-251	160	123	13
-252	162	122	13
-253	164	121	13
-254	166	120	12
-255	167	119	12
