--- src/main/resources/fiji/plugin/trackmate/gui/displaysettings/luts/Solar.lut
@@ -1,257 +1,0 @@
-Index	Red	Green	Blue
-0	51	19	23
-1	52	20	24
-2	53	20	24
-3	55	21	25
-4	56	21	25
-5	57	21	26
-6	58	22	26
-7	59	22	27
-8	61	23	27
-9	62	23	28
-10	63	23	28
-11	64	24	28
-12	66	24	29
-13	67	25	29
-14	68	25	30
-15	69	25	30
-16	71	26	30
-17	72	26	31
-18	73	27	31
-19	74	27	31
-20	76	27	32
-21	77	28	32
-22	78	28	32
-23	79	28	33
-24	81	29	33
-25	82	29	33
-26	83	29	34
-27	84	30	34
-28	86	30	34
-29	87	30	34
-30	88	31	35
-31	89	31	35
-32	91	31	35
-33	92	32	35
-34	93	32	35
-35	94	32	35
-36	96	33	36
-37	97	33	36
-38	98	33	36
-39	99	34	36
-40	101	34	36
-41	102	34	36
-42	103	35	36
-43	104	35	36
-44	106	36	36
-45	107	36	36
-46	108	36	36
-47	109	37	36
-48	111	37	36
-49	112	37	36
-50	113	38	36
-51	114	38	36
-52	116	38	36
-53	117	39	36
-54	118	39	36
-55	119	40	36
-56	120	40	35
-57	122	41	35
-58	123	41	35
-59	124	42	35
-60	125	42	35
-61	126	43	34
-62	127	43	34
-63	129	44	34
-64	130	44	34
-65	131	45	33
-66	132	45	33
-67	133	46	33
-68	134	46	33
-69	135	47	32
-70	136	48	32
-71	137	48	32
-72	138	49	31
-73	139	50	31
-74	140	50	31
-75	141	51	31
-76	142	52	30
-77	143	52	30
-78	144	53	30
-79	145	54	29
-80	146	55	29
-81	147	56	29
-82	148	56	28
-83	149	57	28
-84	150	58	28
-85	151	59	27
-86	151	60	27
-87	152	60	27
-88	153	61	27
-89	154	62	26
-90	155	63	26
-91	156	64	26
-92	156	65	25
-93	157	66	25
-94	158	67	25
-95	159	67	25
-96	160	68	24
-97	160	69	24
-98	161	70	24
-99	162	71	23
-100	163	72	23
-101	164	73	23
-102	164	74	23
-103	165	75	23
-104	166	76	22
-105	166	77	22
-106	167	78	22
-107	168	79	22
-108	169	80	21
-109	169	81	21
-110	170	82	21
-111	171	83	21
-112	171	84	21
-113	172	85	20
-114	173	86	20
-115	173	87	20
-116	174	88	20
-117	175	89	20
-118	175	90	20
-119	176	91	20
-120	177	92	19
-121	177	93	19
-122	178	94	19
-123	179	95	19
-124	179	96	19
-125	180	97	19
-126	180	98	19
-127	181	99	19
-128	182	100	19
-129	182	101	19
-130	183	102	19
-131	183	103	18
-132	184	104	18
-133	185	105	18
-134	185	106	18
-135	186	107	18
-136	186	108	18
-137	187	109	18
-138	187	110	18
-139	188	111	19
-140	188	113	19
-141	189	114	19
-142	190	115	19
-143	190	116	19
-144	191	117	19
-145	191	118	19
-146	192	119	19
-147	192	120	19
-148	193	121	19
-149	193	122	20
-150	194	123	20
-151	194	124	20
-152	195	126	20
-153	195	127	20
-154	196	128	20
-155	196	129	21
-156	197	130	21
-157	197	131	21
-158	198	132	21
-159	198	133	22
-160	199	134	22
-161	199	135	22
-162	199	137	22
-163	200	138	23
-164	200	139	23
-165	201	140	23
-166	201	141	24
-167	202	142	24
-168	202	143	24
-169	203	144	25
-170	203	146	25
-171	203	147	25
-172	204	148	26
-173	204	149	26
-174	205	150	27
-175	205	151	27
-176	205	153	27
-177	206	154	28
-178	206	155	28
-179	207	156	29
-180	207	157	29
-181	207	158	30
-182	208	159	30
-183	208	161	31
-184	209	162	31
-185	209	163	32
-186	209	164	32
-187	210	165	32
-188	210	167	33
-189	210	168	33
-190	211	169	34
-191	211	170	34
-192	211	171	35
-193	212	173	36
-194	212	174	36
-195	212	175	37
-196	213	176	37
-197	213	177	38
-198	213	179	38
-199	214	180	39
-200	214	181	39
-201	214	182	40
-202	215	183	40
-203	215	185	41
-204	215	186	42
-205	216	187	42
-206	216	188	43
-207	216	190	43
-208	216	191	44
-209	217	192	44
-210	217	193	45
-211	217	195	46
-212	217	196	46
-213	218	197	47
-214	218	198	47
-215	218	200	48
-216	218	201	49
-217	219	202	49
-218	219	203	50
-219	219	205	51
-220	219	206	51
-221	220	207	52
-222	220	208	52
-223	220	210	53
-224	220	211	54
-225	220	212	54
-226	221	214	55
-227	221	215	56
-228	221	216	56
-229	221	218	57
-230	221	219	58
-231	222	220	58
-232	222	222	59
-233	222	223	60
-234	222	224	60
-235	222	226	61
-236	222	227	61
-237	223	228	62
-238	223	230	63
-239	223	231	63
-240	223	232	64
-241	223	234	65
-242	223	235	65
-243	223	236	66
-244	223	238	67
-245	223	239	68
-246	224	240	68
-247	224	242	69
-248	224	243	70
-249	224	245	70
-250	224	246	71
-251	224	247	72
-252	224	249	72
-253	224	250	73
-254	224	252	74
-255	224	253	74
