--- src/main/resources/fiji/plugin/trackmate/gui/displaysettings/luts/Speed.lut
@@ -1,257 +1,0 @@
-Index	Red	Green	Blue
-0	23	35	18
-1	23	35	19
-2	23	36	20
-3	23	37	21
-4	23	38	21
-5	24	39	22
-6	24	40	23
-7	24	41	24
-8	24	42	24
-9	24	43	25
-10	24	44	26
-11	25	45	26
-12	25	46	27
-13	25	47	28
-14	25	48	28
-15	25	49	29
-16	25	50	30
-17	25	51	30
-18	25	52	31
-19	25	52	32
-20	25	53	32
-21	25	54	33
-22	25	55	33
-23	25	56	34
-24	24	57	34
-25	24	58	35
-26	24	59	35
-27	24	60	36
-28	24	61	36
-29	24	62	37
-30	23	63	37
-31	23	64	38
-32	23	65	38
-33	23	66	39
-34	22	67	39
-35	22	68	39
-36	22	69	40
-37	21	70	40
-38	21	71	40
-39	21	72	41
-40	20	73	41
-41	20	74	41
-42	20	74	42
-43	19	75	42
-44	19	76	42
-45	18	77	42
-46	18	78	43
-47	17	79	43
-48	17	80	43
-49	16	81	43
-50	16	82	43
-51	15	83	44
-52	15	84	44
-53	15	85	44
-54	14	86	44
-55	14	87	44
-56	13	88	44
-57	13	89	44
-58	12	90	44
-59	12	91	44
-60	12	92	44
-61	11	92	44
-62	11	93	44
-63	11	94	44
-64	11	95	44
-65	10	96	44
-66	10	97	44
-67	10	98	44
-68	10	99	44
-69	11	100	44
-70	11	101	44
-71	11	102	43
-72	12	103	43
-73	12	103	43
-74	13	104	43
-75	13	105	43
-76	14	106	42
-77	15	107	42
-78	16	108	42
-79	17	109	41
-80	18	110	41
-81	19	111	41
-82	20	111	41
-83	21	112	40
-84	22	113	40
-85	24	114	39
-86	25	115	39
-87	26	116	39
-88	28	117	38
-89	29	117	38
-90	31	118	37
-91	32	119	37
-92	34	120	36
-93	35	121	36
-94	37	121	35
-95	38	122	35
-96	40	123	34
-97	42	124	33
-98	43	125	33
-99	45	125	32
-100	46	126	32
-101	48	127	31
-102	50	128	30
-103	52	129	30
-104	53	129	29
-105	55	130	28
-106	57	131	28
-107	59	131	27
-108	60	132	26
-109	62	133	26
-110	64	134	25
-111	66	134	24
-112	67	135	23
-113	69	136	23
-114	71	137	22
-115	73	137	21
-116	75	138	20
-117	77	139	20
-118	78	139	19
-119	80	140	18
-120	82	141	17
-121	84	141	16
-122	86	142	16
-123	87	143	15
-124	89	143	14
-125	91	144	13
-126	93	145	12
-127	95	145	12
-128	97	146	11
-129	98	146	10
-130	100	147	9
-131	102	148	9
-132	104	148	8
-133	106	149	8
-134	107	150	7
-135	109	150	7
-136	111	151	6
-137	113	151	6
-138	115	152	5
-139	116	153	5
-140	118	153	5
-141	120	154	5
-142	122	155	5
-143	124	155	5
-144	125	156	5
-145	127	156	6
-146	129	157	6
-147	131	158	6
-148	132	158	7
-149	134	159	7
-150	136	159	8
-151	138	160	9
-152	139	161	10
-153	141	161	11
-154	143	162	12
-155	145	162	13
-156	146	163	14
-157	148	164	15
-158	150	164	16
-159	151	165	17
-160	153	165	18
-161	155	166	20
-162	157	167	21
-163	158	167	22
-164	160	168	23
-165	162	168	25
-166	163	169	26
-167	165	170	28
-168	167	170	29
-169	168	171	30
-170	170	171	32
-171	171	172	33
-172	173	173	35
-173	175	173	36
-174	176	174	38
-175	178	175	40
-176	179	175	41
-177	181	176	43
-178	182	177	44
-179	184	177	46
-180	185	178	48
-181	187	179	50
-182	188	179	51
-183	190	180	53
-184	191	181	55
-185	193	181	57
-186	194	182	59
-187	195	183	60
-188	197	183	62
-189	198	184	64
-190	199	185	66
-191	201	186	68
-192	202	186	70
-193	203	187	72
-194	205	188	74
-195	206	189	76
-196	207	190	78
-197	208	190	80
-198	209	191	82
-199	211	192	84
-200	212	193	86
-201	213	194	88
-202	214	195	91
-203	215	196	93
-204	216	196	95
-205	217	197	97
-206	218	198	99
-207	219	199	101
-208	220	200	103
-209	221	201	106
-210	222	202	108
-211	223	203	110
-212	223	204	112
-213	224	205	114
-214	225	206	116
-215	226	207	119
-216	227	208	121
-217	228	209	123
-218	229	210	125
-219	229	211	127
-220	230	212	129
-221	231	213	132
-222	232	214	134
-223	232	215	136
-224	233	216	138
-225	234	217	140
-226	235	219	143
-227	235	220	145
-228	236	221	147
-229	237	222	149
-230	238	223	151
-231	238	224	153
-232	239	225	156
-233	240	226	158
-234	240	227	160
-235	241	229	162
-236	242	230	164
-237	242	231	166
-238	243	232	168
-239	244	233	171
-240	244	234	173
-241	245	236	175
-242	246	237	177
-243	246	238	179
-244	247	239	181
-245	248	240	183
-246	248	241	185
-247	249	243	188
-248	250	244	190
-249	250	245	192
-250	251	246	194
-251	252	247	196
-252	252	249	198
-253	253	250	200
-254	254	251	202
-255	254	252	205
