--- src/main/resources/fiji/plugin/trackmate/gui/displaysettings/luts/Tempo.lut
@@ -1,257 +1,0 @@
-Index	Red	Green	Blue
-0	20	29	67
-1	21	30	68
-2	21	31	68
-3	21	32	69
-4	21	33	69
-5	22	34	70
-6	22	35	70
-7	22	36	71
-8	22	37	71
-9	22	38	72
-10	23	39	72
-11	23	40	73
-12	23	41	74
-13	23	42	74
-14	24	43	75
-15	24	44	75
-16	24	45	76
-17	24	46	76
-18	24	47	77
-19	24	47	78
-20	25	48	78
-21	25	49	79
-22	25	50	79
-23	25	51	80
-24	25	52	80
-25	26	53	81
-26	26	54	82
-27	26	55	82
-28	26	56	83
-29	26	57	83
-30	26	58	84
-31	26	59	84
-32	27	60	85
-33	27	60	86
-34	27	61	86
-35	27	62	87
-36	27	63	87
-37	27	64	88
-38	27	65	89
-39	27	66	89
-40	27	67	90
-41	27	68	90
-42	27	69	91
-43	28	70	91
-44	28	70	92
-45	28	71	93
-46	28	72	93
-47	28	73	94
-48	28	74	94
-49	28	75	95
-50	28	76	96
-51	28	77	96
-52	28	78	97
-53	28	79	97
-54	27	80	98
-55	27	81	98
-56	27	81	99
-57	27	82	100
-58	27	83	100
-59	27	84	101
-60	27	85	101
-61	27	86	102
-62	27	87	102
-63	27	88	103
-64	26	89	103
-65	26	90	104
-66	26	91	105
-67	26	92	105
-68	26	93	106
-69	25	94	106
-70	25	94	107
-71	25	95	107
-72	25	96	108
-73	24	97	108
-74	24	98	109
-75	24	99	109
-76	23	100	110
-77	23	101	110
-78	23	102	111
-79	22	103	111
-80	22	104	112
-81	22	105	112
-82	21	106	113
-83	21	107	113
-84	21	108	114
-85	20	109	114
-86	20	109	115
-87	20	110	115
-88	19	111	115
-89	19	112	116
-90	19	113	116
-91	18	114	117
-92	18	115	117
-93	18	116	117
-94	17	117	118
-95	17	118	118
-96	17	119	119
-97	17	120	119
-98	16	121	119
-99	16	122	120
-100	16	123	120
-101	16	124	120
-102	16	125	121
-103	16	125	121
-104	17	126	121
-105	17	127	122
-106	17	128	122
-107	18	129	122
-108	18	130	123
-109	19	131	123
-110	19	132	123
-111	20	133	123
-112	21	134	124
-113	22	135	124
-114	23	136	124
-115	24	137	124
-116	25	137	125
-117	26	138	125
-118	28	139	125
-119	29	140	125
-120	31	141	125
-121	32	142	126
-122	34	143	126
-123	35	144	126
-124	37	144	126
-125	38	145	126
-126	40	146	127
-127	42	147	127
-128	44	148	127
-129	46	149	127
-130	47	149	127
-131	49	150	128
-132	51	151	128
-133	53	152	128
-134	55	153	128
-135	57	153	128
-136	59	154	129
-137	61	155	129
-138	63	156	129
-139	65	157	129
-140	67	157	129
-141	69	158	130
-142	71	159	130
-143	73	160	130
-144	75	160	131
-145	77	161	131
-146	79	162	131
-147	81	162	131
-148	83	163	132
-149	85	164	132
-150	87	165	133
-151	89	165	133
-152	91	166	133
-153	93	167	134
-154	95	167	134
-155	97	168	135
-156	99	169	135
-157	101	169	136
-158	103	170	136
-159	105	171	137
-160	106	171	137
-161	108	172	138
-162	110	173	138
-163	112	174	139
-164	114	174	139
-165	116	175	140
-166	118	176	141
-167	119	176	141
-168	121	177	142
-169	123	178	143
-170	125	178	143
-171	127	179	144
-172	128	180	145
-173	130	180	145
-174	132	181	146
-175	134	182	147
-176	136	182	148
-177	137	183	148
-178	139	184	149
-179	141	184	150
-180	142	185	151
-181	144	186	152
-182	146	186	153
-183	148	187	153
-184	149	188	154
-185	151	188	155
-186	153	189	156
-187	154	190	157
-188	156	191	158
-189	157	191	159
-190	159	192	160
-191	161	193	161
-192	162	193	162
-193	164	194	163
-194	166	195	164
-195	167	195	165
-196	169	196	166
-197	170	197	167
-198	172	198	168
-199	174	198	169
-200	175	199	170
-201	177	200	171
-202	178	201	172
-203	180	201	174
-204	181	202	175
-205	183	203	176
-206	184	204	177
-207	186	204	178
-208	187	205	179
-209	189	206	181
-210	190	207	182
-211	192	207	183
-212	193	208	184
-213	195	209	185
-214	196	210	187
-215	198	210	188
-216	199	211	189
-217	201	212	190
-218	202	213	192
-219	204	214	193
-220	205	214	194
-221	207	215	195
-222	208	216	197
-223	210	217	198
-224	211	218	199
-225	212	218	201
-226	214	219	202
-227	215	220	203
-228	217	221	205
-229	218	222	206
-230	220	223	207
-231	221	223	209
-232	222	224	210
-233	224	225	212
-234	225	226	213
-235	227	227	214
-236	228	228	216
-237	229	229	217
-238	231	229	219
-239	232	230	220
-240	234	231	221
-241	235	232	223
-242	236	233	224
-243	238	234	226
-244	239	235	227
-245	241	236	229
-246	242	237	230
-247	243	238	232
-248	245	238	233
-249	246	239	235
-250	247	240	236
-251	249	241	238
-252	250	242	239
-253	252	243	241
-254	253	244	242
-255	254	245	244
