--- src/main/resources/fiji/plugin/trackmate/gui/displaysettings/luts/Thermal.lut
@@ -1,257 +1,0 @@
-Index	Red	Green	Blue
-0	3	35	51
-1	4	35	53
-2	4	36	55
-3	4	37	57
-4	4	38	58
-5	4	38	60
-6	5	39	62
-7	5	40	64
-8	5	40	66
-9	5	41	68
-10	5	42	70
-11	6	42	73
-12	6	43	75
-13	6	43	77
-14	7	44	79
-15	7	44	81
-16	8	45	83
-17	8	46	86
-18	9	46	88
-19	10	47	90
-20	10	47	93
-21	11	48	95
-22	12	48	98
-23	13	48	100
-24	14	49	103
-25	15	49	105
-26	16	50	108
-27	17	50	110
-28	18	50	113
-29	20	50	116
-30	21	51	118
-31	23	51	121
-32	24	51	124
-33	26	51	126
-34	27	51	129
-35	29	51	132
-36	31	51	134
-37	33	51	137
-38	35	51	140
-39	37	51	142
-40	39	51	144
-41	41	51	147
-42	43	51	149
-43	46	51	151
-44	48	51	153
-45	50	51	154
-46	53	50	156
-47	55	51	157
-48	57	51	157
-49	59	51	158
-50	61	51	159
-51	63	51	159
-52	65	52	159
-53	67	52	159
-54	69	53	159
-55	71	53	159
-56	73	54	159
-57	74	54	158
-58	76	55	158
-59	78	55	158
-60	79	56	157
-61	81	57	157
-62	83	57	156
-63	84	58	156
-64	86	59	156
-65	87	59	155
-66	89	60	155
-67	90	61	154
-68	92	61	154
-69	93	62	153
-70	95	63	153
-71	96	64	152
-72	97	64	152
-73	99	65	151
-74	100	66	151
-75	102	66	150
-76	103	67	150
-77	104	67	149
-78	106	68	149
-79	107	69	148
-80	108	69	148
-81	110	70	148
-82	111	71	147
-83	112	71	147
-84	114	72	146
-85	115	73	146
-86	116	73	146
-87	118	74	145
-88	119	74	145
-89	121	75	144
-90	122	76	144
-91	123	76	144
-92	125	77	143
-93	126	77	143
-94	127	78	143
-95	129	78	142
-96	130	79	142
-97	131	80	142
-98	133	80	142
-99	134	81	141
-100	135	81	141
-101	137	82	141
-102	138	82	140
-103	140	83	140
-104	141	83	140
-105	142	84	139
-106	144	84	139
-107	145	85	139
-108	147	85	138
-109	148	86	138
-110	150	86	138
-111	151	87	137
-112	152	87	137
-113	154	88	136
-114	155	88	136
-115	157	89	136
-116	158	89	135
-117	160	90	135
-118	161	90	134
-119	163	91	134
-120	164	91	133
-121	166	92	133
-122	167	92	133
-123	169	93	132
-124	170	93	131
-125	172	93	131
-126	173	94	130
-127	175	94	130
-128	176	95	129
-129	178	95	129
-130	179	96	128
-131	181	96	127
-132	182	97	127
-133	184	97	126
-134	185	97	125
-135	187	98	124
-136	188	98	124
-137	190	99	123
-138	191	99	122
-139	193	100	121
-140	194	100	120
-141	196	101	120
-142	197	101	119
-143	199	102	118
-144	200	102	117
-145	202	103	116
-146	203	103	115
-147	205	104	114
-148	206	104	113
-149	207	105	112
-150	209	105	111
-151	210	106	110
-152	212	107	109
-153	213	107	108
-154	214	108	107
-155	216	108	105
-156	217	109	104
-157	219	110	103
-158	220	110	102
-159	221	111	101
-160	222	112	100
-161	224	112	98
-162	225	113	97
-163	226	114	96
-164	227	115	95
-165	229	116	93
-166	230	116	92
-167	231	117	91
-168	232	118	90
-169	233	119	88
-170	234	120	87
-171	235	121	86
-172	236	122	85
-173	237	123	83
-174	238	124	82
-175	239	125	81
-176	240	126	80
-177	241	127	79
-178	241	128	78
-179	242	130	76
-180	243	131	75
-181	243	132	74
-182	244	133	73
-183	245	135	72
-184	245	136	71
-185	246	137	70
-186	246	139	69
-187	247	140	69
-188	247	141	68
-189	248	143	67
-190	248	144	66
-191	249	146	65
-192	249	147	65
-193	249	148	64
-194	250	150	63
-195	250	151	63
-196	250	153	62
-197	250	154	62
-198	250	156	62
-199	251	157	61
-200	251	159	61
-201	251	161	61
-202	251	162	60
-203	251	164	60
-204	251	165	60
-205	251	167	60
-206	251	168	60
-207	251	170	60
-208	251	172	60
-209	251	173	60
-210	251	175	60
-211	251	176	60
-212	251	178	60
-213	251	180	60
-214	251	181	61
-215	251	183	61
-216	250	184	61
-217	250	186	62
-218	250	188	62
-219	250	189	62
-220	250	191	63
-221	249	193	63
-222	249	194	64
-223	249	196	64
-224	249	198	65
-225	248	199	65
-226	248	201	66
-227	248	203	67
-228	247	204	67
-229	247	206	68
-230	246	208	69
-231	246	209	69
-232	246	211	70
-233	245	213	71
-234	245	214	72
-235	244	216	72
-236	244	218	73
-237	243	219	74
-238	243	221	75
-239	242	223	76
-240	242	224	77
-241	241	226	77
-242	241	228	78
-243	240	229	79
-244	239	231	80
-245	239	233	81
-246	238	234	82
-247	237	236	83
-248	237	238	84
-249	236	240	85
-250	235	241	86
-251	235	243	86
-252	234	245	87
-253	233	246	88
-254	232	248	89
-255	231	250	90
