--- src/main/resources/fiji/plugin/trackmate/gui/displaysettings/luts/Turbid.lut
@@ -1,257 +1,0 @@
-Index	Red	Green	Blue
-0	34	30	27
-1	35	31	27
-2	36	31	28
-3	37	32	28
-4	38	33	29
-5	39	33	29
-6	40	34	30
-7	41	35	30
-8	42	35	31
-9	43	36	31
-10	44	37	31
-11	45	37	32
-12	46	38	32
-13	47	39	33
-14	48	39	33
-15	49	40	34
-16	50	40	34
-17	51	41	35
-18	52	42	35
-19	53	42	36
-20	54	43	36
-21	55	44	36
-22	56	44	37
-23	57	45	37
-24	58	45	38
-25	59	46	38
-26	60	47	39
-27	61	47	39
-28	62	48	39
-29	63	49	40
-30	64	49	40
-31	65	50	41
-32	66	50	41
-33	67	51	41
-34	68	52	42
-35	69	52	42
-36	70	53	42
-37	71	53	43
-38	72	54	43
-39	73	55	44
-40	74	55	44
-41	75	56	44
-42	76	56	45
-43	77	57	45
-44	78	58	45
-45	79	58	46
-46	80	59	46
-47	81	59	46
-48	82	60	47
-49	83	60	47
-50	84	61	47
-51	85	62	48
-52	86	62	48
-53	87	63	48
-54	88	63	49
-55	89	64	49
-56	90	65	49
-57	92	65	49
-58	93	66	50
-59	94	66	50
-60	95	67	50
-61	96	68	51
-62	97	68	51
-63	98	69	51
-64	99	69	51
-65	100	70	52
-66	101	71	52
-67	102	71	52
-68	103	72	52
-69	104	72	53
-70	105	73	53
-71	106	74	53
-72	107	74	53
-73	108	75	53
-74	109	75	54
-75	110	76	54
-76	111	76	54
-77	112	77	54
-78	113	78	54
-79	114	78	55
-80	115	79	55
-81	116	80	55
-82	117	80	55
-83	119	81	55
-84	120	81	55
-85	121	82	55
-86	122	83	56
-87	123	83	56
-88	124	84	56
-89	125	85	56
-90	126	85	56
-91	127	86	56
-92	128	86	56
-93	129	87	56
-94	130	88	57
-95	131	88	57
-96	132	89	57
-97	133	90	57
-98	134	90	57
-99	135	91	57
-100	136	92	57
-101	137	92	57
-102	138	93	57
-103	139	94	57
-104	140	94	57
-105	141	95	57
-106	142	96	57
-107	142	97	58
-108	143	97	58
-109	144	98	58
-110	145	99	58
-111	146	99	58
-112	147	100	58
-113	148	101	58
-114	149	102	58
-115	150	102	58
-116	151	103	58
-117	152	104	58
-118	153	105	58
-119	154	105	58
-120	154	106	58
-121	155	107	58
-122	156	108	58
-123	157	108	58
-124	158	109	58
-125	159	110	59
-126	160	111	59
-127	161	111	59
-128	161	112	59
-129	162	113	59
-130	163	114	59
-131	164	115	59
-132	165	116	59
-133	166	116	59
-134	166	117	59
-135	167	118	59
-136	168	119	60
-137	169	120	60
-138	169	121	60
-139	170	121	60
-140	171	122	60
-141	172	123	60
-142	173	124	60
-143	173	125	61
-144	174	126	61
-145	175	127	61
-146	175	128	61
-147	176	129	61
-148	177	129	62
-149	178	130	62
-150	178	131	62
-151	179	132	62
-152	180	133	63
-153	180	134	63
-154	181	135	63
-155	182	136	64
-156	182	137	64
-157	183	138	64
-158	184	139	65
-159	184	140	65
-160	185	141	66
-161	185	142	66
-162	186	143	66
-163	187	144	67
-164	187	145	67
-165	188	146	68
-166	188	147	68
-167	189	148	69
-168	190	149	69
-169	190	150	70
-170	191	151	71
-171	191	152	71
-172	192	153	72
-173	192	154	73
-174	193	155	73
-175	193	156	74
-176	194	157	75
-177	194	158	75
-178	195	159	76
-179	195	160	77
-180	196	161	78
-181	196	162	78
-182	197	163	79
-183	197	164	80
-184	198	165	81
-185	198	167	82
-186	199	168	83
-187	199	169	84
-188	200	170	84
-189	200	171	85
-190	201	172	86
-191	201	173	87
-192	202	174	88
-193	202	175	89
-194	202	176	90
-195	203	177	91
-196	203	178	92
-197	204	180	94
-198	204	181	95
-199	205	182	96
-200	205	183	97
-201	206	184	98
-202	206	185	99
-203	206	186	100
-204	207	187	101
-205	207	188	102
-206	208	189	104
-207	208	191	105
-208	209	192	106
-209	209	193	107
-210	210	194	108
-211	210	195	110
-212	210	196	111
-213	211	197	112
-214	211	198	113
-215	212	199	115
-216	212	201	116
-217	213	202	117
-218	213	203	119
-219	214	204	120
-220	214	205	121
-221	215	206	122
-222	215	207	124
-223	216	208	125
-224	216	210	126
-225	217	211	128
-226	217	212	129
-227	217	213	130
-228	218	214	132
-229	218	215	133
-230	219	216	135
-231	219	218	136
-232	220	219	137
-233	220	220	139
-234	221	221	140
-235	221	222	142
-236	222	223	143
-237	222	224	144
-238	223	226	146
-239	223	227	147
-240	224	228	149
-241	225	229	150
-242	225	230	152
-243	226	231	153
-244	226	232	155
-245	227	234	156
-246	227	235	157
-247	228	236	159
-248	228	237	160
-249	229	238	162
-250	229	239	163
-251	230	241	165
-252	231	242	166
-253	231	243	168
-254	232	244	169
-255	232	245	171
